% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\docType{methods}
\name{addCellMeta2cellexalvr}
\alias{addCellMeta2cellexalvr}
\alias{addCellMeta2cellexalvr,cellexalvrR-method}
\title{description of function addCellMeta2cellexalvr}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cell.meta}{A matrix of cell metadata}
}
\description{
Adds per cell metadata to a cellexalvrObj
}
