% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\name{addCellMeta2cellexalvr}
\alias{addCellMeta2cellexalvr}
\title{Adds per cell metadata  to a cellexalvrObj}
\usage{
addCellMeta2cellexalvr(cellexalObj, cell.meta)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cell.meta}{A matrix of cell metadata}
}
\description{
Adds per cell metadata  to a cellexalvrObj
}
