% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\docType{methods}
\name{addDRC2cellexalvr}
\alias{addDRC2cellexalvr}
\alias{addDRC2cellexalvr,cellexalvrR-method}
\title{description of function 'addDRC2cellexalvr'}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{drcmatrix}{A (3 columns) matrix of coordinates}

\item{name}{A name for the object (default = graph<n>)}
}
\description{
Adds drc coordinates to a 'cellexalvrObj'
}
