% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\docType{methods}
\name{addDRC2cellexalvr}
\alias{addDRC2cellexalvr}
\alias{addDRC2cellexalvr,cellexalvrR-method}
\title{Adds drc coordinates to a 'cellexalvrObj'}
\usage{
addDRC2cellexalvr(cellexalObj, drcmatrix, name = NULL)

\S4method{addDRC2cellexalvr}{cellexalvrR}(cellexalObj, drcmatrix, name = NULL)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{drcmatrix}{A (3 columns) matrix of coordinates}

\item{name}{A name for the object (default = graph<n>)}
}
\description{
Adds drc coordinates to a 'cellexalvrObj'
}
\details{
addDRC2cellexalvr is a helper function that applies some tests 
of usabilty to a 3D DRC matrix object and adds it to the cellexalvrR object.
}
