% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\docType{methods}
\name{addFACS2cellexalvr}
\alias{addFACS2cellexalvr}
\alias{addFACS2cellexalvr,cellexalvrR-method}
\title{description of function addFACS2cellexalvr}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{facs.data}{A matrix of surface marker intensities}
}
\description{
Adds FACS index to a cellexalvrObj
}
