% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\name{addFACS2cellexalvr}
\alias{addFACS2cellexalvr}
\title{Adds FACS index to a cellexalvrObj}
\usage{
addFACS2cellexalvr(cellexalObj, facs.data)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{facs.data}{A matrix of surface marker intensities}
}
\description{
Adds FACS index to a cellexalvrObj
}
