% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\docType{methods}
\name{addMDS2cellexalvr}
\alias{addMDS2cellexalvr}
\alias{addMDS2cellexalvr,cellexalvrR-method}
\title{description of function addMDS2cellexalvr}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{mdsmatrix}{A matrix of coordinates}

\item{name}{A name for the object (default = graph<n>)}
}
\description{
Adds mds coordinates to a cellexalvrObj
}
