% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrExportFunctions.R
\name{addMDS2cellexalvr}
\alias{addMDS2cellexalvr}
\title{Adds mds coordinates to a cellexalvrObj}
\usage{
addMDS2cellexalvr(cellexalObj, mdsmatrix)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}
}
\description{
Adds mds coordinates to a cellexalvrObj
}
