% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\docType{methods}
\name{addNewVelocity}
\alias{addNewVelocity}
\alias{addNewVelocity,cellexalvrR-method}
\title{description of function 'addNewVelocity'}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{drcmatrix}{A (3 columns) matrix of coordinates}

\item{dr.name}{A name for the object (default = graph<n>)}
}
\description{
Adds drc coordinates to a 'cellexalvrObj'
}
\details{
addNewVelocity is a simple helper function that applies some tests 
of usabilty to a 3D DRC matrix object and adds it to the cellexalvrR object.
}
