% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{addSelection,cellexalTime,cellexalvrR-method}
\alias{addSelection,cellexalTime,cellexalvrR-method}
\title{Method to copy a selection from cellexalObj to cellexalObj.
This function makes sure that the linear grouping is fitting into the object.}
\usage{
\S4method{addSelection}{cellexalTime,cellexalvrR}(x, cellexalObj, upstreamSelection = NULL)
}
\description{
Method to copy a selection from cellexalObj to cellexalObj.
This function makes sure that the linear grouping is fitting into the object.
}
