% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{addSelection}
\alias{addSelection}
\alias{addSelection,cellexalTime-method}
\title{description of function addSelection}
\arguments{
\item{x}{the cellexalTime object}

\item{cellexalObj}{the cellexalvrR object to add the time as group}

\item{upstreamSelection}{which group was the basis for this timeline?}
}
\description{
add this selection to any cellexal object
}
