% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{addSelection}
\alias{addSelection}
\alias{addSelection,cellexalTime,cellexalvrR-method}
\title{savely add a cellexalTime object to a cellexalvR object}
\usage{
addSelection(x, cellexalObj, upstreamSelection = NULL)

\S4method{addSelection}{cellexalTime,cellexalvrR}(x, cellexalObj, upstreamSelection = NULL)
}
\arguments{
\item{x}{the cellexalTime object}

\item{cellexalObj}{the cellexalvrR object to add the time as group}

\item{upstreamSelection}{which group was the basis for this timeline?}
}
\description{
add this cellexalTime to the corresponding cellexal object
}
\details{
Method to copy a selection from cellexalObj to cellexalObj.
This function makes sure that the linear grouping is fitting into the object.
}
