% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\docType{methods}
\name{addVelocityToExistingDR}
\alias{addVelocityToExistingDR}
\alias{addVelocityToExistingDR,cellexalvrR-method}
\title{Adds RNA volocity data to an existing DR methods in an 'cellexalvrObj' object}
\usage{
addVelocityToExistingDR(cellexalObj, velo.arrows, dr.name)

\S4method{addVelocityToExistingDR}{cellexalvrR}(cellexalObj, velo.arrows, dr.name)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{velo.arrows}{A 6 column matrix describing RNA velocity arrows}

\item{dr.name}{The name of the DR coordinates being added to}
}
\description{
Adds RNA volocity data to an existing DR methods in an 'cellexalvrObj' object.
}
\details{
This function adds RNA velocity coordinates to an already existing set of dimension reduction coordinates.
}
