% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_BioData.R
\docType{methods}
\name{as_BioData}
\alias{as_BioData}
\alias{as_BioData,cellexalvr-method}
\title{description of function as}
\usage{
as_BioData(dat)
}
\arguments{
\item{dat}{the cellexalvr object}
}
\description{
create a BioData object from a cellexalvr object
}
