% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cellexalvrR.R
\docType{methods}
\name{as_cellexalvrR,H5File-method}
\alias{as_cellexalvrR,H5File-method}
\title{Cast an AnnData object to a cellexalvrR object}
\usage{
\S4method{as_cellexalvrR}{H5File}(x, meta.cell.groups = NULL,
  meta.genes.groups = NULL, userGroups = NULL, outpath = getwd(),
  specie)
}
\description{
Cast an AnnData object to a cellexalvrR object
}
