% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cellexalvrR.R
\docType{methods}
\name{as_cellexalvrR,loom-method}
\alias{as_cellexalvrR,loom-method}
\title{Cast a loom object to a cellexalvrR object}
\usage{
\S4method{as_cellexalvrR}{loom}(x, meta.cell.groups = NULL,
  meta.genes.groups = NULL, userGroups = NULL, outpath = getwd(),
  specie)
}
\description{
Cast a loom object to a cellexalvrR object
}
