% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cellexalvrR.R
\docType{methods}
\name{as_cellexalvrR}
\alias{as_cellexalvrR}
\alias{as_cellexalvrR,environment-method}
\title{convert a BioData object to cellexalvrR keeping all 3D mds objects.}
\usage{
as_cellexalvrR(x, meta.cell.groups, meta.genes.groups = NULL,
  userGroups = NULL, outpath = getwd(), specie)
}
\arguments{
\item{x}{the BioData 'object'}

\item{meta.cell.groups}{which x$samples column to convert to meta.cell classes}

\item{meta.genes.groups}{which annotation columns to keep (default NULL)}

\item{userGroups}{which x§samples columns to add to the userGroups slot}

\item{outpath}{set the outpath of the object (default getwd())}

\item{specie}{set the specie to either mouse or human (default check gene names)}
}
\description{
convert a BioData list (BioData libraray not loaded) into a cellexalvrR obejct
}
