% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.name.R
\docType{methods}
\name{auto_name}
\alias{auto_name}
\alias{auto_name,cellexalvr-method}
\title{description of function auto.name}
\usage{
auto_name(cellexalObj, new = TRUE)
}
\arguments{
\item{cellexalObj}{the object}

\item{new}{want the last name or a new name default=TRUE (new)}
}
\description{
This is a simple (internal) function to return a new name if the user has not given a name for a grouing.
}
