% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchPointAnalysis.R
\docType{methods}
\name{branch.point.analysis}
\alias{branch.point.analysis}
\alias{branch.point.analysis,character-method}
\title{simple preload cellexalObj}
\usage{
\S4method{branch.point.analysis}{character}(cellexalObj, cellidfile,
  maxsig, outfile)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{maxsig}{The number of differentials to be returned}

\item{outfile}{The name of the output file}
}
\description{
simple preload the cellexalObj and run the cellexalvrR function afterwards
}
\keyword{analysis}
\keyword{branch}
\keyword{point}
