% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_vignette.R
\docType{methods}
\name{buildVignette}
\alias{buildVignette}
\title{helper to prepare the package for github}
\usage{
buildVignette(dir = ".")
}
\arguments{
\item{dir}{the path of the object's dev folder}
}
\description{
Build Vignettes and make them available
}
\details{
Hack to build the vignettes and make them available to devtools::install_github()
Found this fix on https://community.rstudio.com/t/browsevignettes-mypackage-saying-no-vignettes-found/68656/6
}
