% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellCycle}
\alias{CellCycle}
\title{A simple table containing the Human and mouse orthologe CellCycle genes from PMID17994010
"Genome-scale RNAi profiling of cell division in human tissue culture cells."}
\format{data.frame}
\usage{
CellCycle
}
\description{
The data can be used by stating
useInbuiltGOIlists (cellexalObj, 'CellCycle' ) #function definition in file 'useInbuiltGOIlists.R'
And it is used to visualize the cell cycle genes in the VR environment.
Only the genes also identifyable in mouse were used here.
}
\keyword{data}
