% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03cellexalGrouping.R
\docType{class}
\name{cellexalGrouping-class}
\alias{cellexalGrouping-class}
\title{cellexalGrouping class definition}
\description{
A simple wrapper to hadle the time and time color mappings.
}
\details{
the cellexalGrouping class is mainly to bring a usable order into the time mess
}
\section{Slots}{

\describe{
\item{\code{gname}}{the group name}

\item{\code{selectionFile}}{the VR selection file that is the basis for this grouping}

\item{\code{order}}{the order the cells have been selected in the VR process}

\item{\code{drc}}{the drc name this object has been selected from}

\item{\code{col}}{the color vector for these groups}

\item{\code{error}}{a string vector that contains all error messages}

\item{\code{timeObj}}{an optional slot to store a cellexalTime object}

\item{\code{heatmapBasename}}{the filename basis for the heatmap related to this grouping}
}}

