% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellexalObj}
\alias{cellexalObj}
\title{The cellexlvrR object described in the vignette}
\format{
cellexalvrR
}
\usage{
cellexalObj
}
\description{
THE test dataset
}
\keyword{data}
