% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{class}
\name{cellexalTime-class}
\alias{cellexalTime-class}
\title{cellexalTime class definition}
\description{
A simple wrapper to hadle the time and time color mappings.
}
\details{
the cellexalTime class is mainly to bring a usable order into the time mess
}
\section{Slots}{

\describe{
\item{\code{dat}}{all data needed for the time plotting and group creation}

\item{\code{gname}}{the group name}

\item{\code{drc}}{the drc name this object has been selected from}

\item{\code{error}}{the error message if a catched not fatal error has occured}

\item{\code{geneClusters}}{a list of gene clusters that are linked to a timeline}
}}

