% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalvrClass.R
\docType{class}
\name{cellexalvr-class}
\alias{cellexalvr-class}
\title{cellexalvr class definition (old)}
\description{
This class definition is totally empty and only incuded here to allow for a smooth update of the old objects.
}
