% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrClass.R
\docType{class}
\name{cellexalvr}
\alias{cellexalvr}
\title{Cellexalvr, a class to interact with the VR environment developed by the Computational Genomics goup in Lund}
\format{\code{\link{R6Class}} object.}
\usage{
cellexalvr
}
\value{
Object of \code{\link{R6Class}} to interact with the VR environment.
}
\description{
Cellexalvr, a class to interact with the VR environment developed by the Computational Genomics goup in Lund
}
\section{Fields}{

\describe{
\item{\code{data}}{the numerical data as data.frame}

\item{\code{meta.cell}}{the sample annotation as data.frame}

\item{\code{meta.gene}}{the row annotation as data.frame}

\item{\code{usedObj}}{a multi purpose list to store whichever ananlyis results do not fit in the stats list}
}}

\examples{
set.seed(1)
dat =  matrix(rnorm(1000),ncol=10) 
colnames(dat) <- paste('Sample', 1:10)
rownames(dat) <- paste( 'gene', 1:100)
samples <- data.frame(SampleID = 1:10, sname = colnames(dat) )
annotation <- data.frame( GeneID = paste( 'gene', 1:100), Start= 101:200 )
x <- cellexalvr( data=dat, meta.gene=annotation, meta.cell = samples )
}
\keyword{cellexalvr}
