% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalvrClass.R
\docType{class}
\name{cellexalvrR}
\alias{cellexalvrR}
\title{cellexalvrR class definition}
\description{
The R backend for the CellexalVR 3D application
}
\section{Slots}{

\describe{
\item{\code{data}}{the expression matrix}

\item{\code{meta.cell}}{the cell level meta information (matrix)}

\item{\code{meta.gene}}{the gene level meta information (matrix)}

\item{\code{userGroups}}{internally used to store the user defined groupings from the 3D process (data.frame)}

\item{\code{colors}}{a list for each userGroups entry defining the color sheme for this grouping}

\item{\code{mds}}{a list of all mds objects to be visible in the 3D application}

\item{\code{index}}{a matrix for FACS or other linear numeric data that should be available for colouring in the 3D application}

\item{\code{tfs}}{depricated not used any more}

\item{\code{specie}}{the species this data is from (mouse or human)}
}}

