% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratExtensions.R
\docType{methods}
\name{changeIdent}
\alias{changeIdent}
\alias{changeIdent,seurat-method}
\title{description of function changeIdent}
\arguments{
\item{seuratObj}{A cellexalvr object}

\item{cell}{A cellexalvr object}

\item{new.idents}{TEXT MISSING}
}
\description{
Sets new cell indentities from a given list
}
\keyword{cell}
\keyword{type}
