% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratExtensions.R
\name{changeIdent}
\alias{changeIdent}
\title{Sets new cell indentities from a given list}
\usage{
changeIdent(seuratObj, new.idents)
}
\arguments{
\item{seuratObj}{A cellexalvr object}

\item{cell}{A cellexalvr object}
}
\description{
Sets new cell indentities from a given list
}
\keyword{cell}
\keyword{type}
