% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{check}
\alias{check}
\alias{check,cellexalvrR-method}
\title{description of function check}
\arguments{
\item{x}{the cellexalTime object}
}
\description{
checks for NA elements in the table and removes them
}
