% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\docType{methods}
\name{check}
\alias{check}
\alias{check,cellexalvrR-method}
\title{cellexal internal checks}
\usage{
check(x, silent = FALSE)

\S4method{check}{cellexalvrR}(x, silent = FALSE)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{silent}{do not report warnings (default FALSE)}
}
\value{
the checked cellexalvrR object
}
\description{
checks cellexalvrR internals
}
\details{
The cellexalvrR sanity check.
Prints out the problems and sets a x@usedObj$ceckPassed boolean value.
This function is called in the export2cellexalvr function and others.
}
