% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkgrouping.R
\name{checkGrouping,cellexalvrR,cellexalGrouping-method}
\alias{checkGrouping,cellexalvrR,cellexalGrouping-method}
\title{Grouping as cellexalGrouping object will directly check this object.}
\usage{
\S4method{checkGrouping}{cellexalvrR,cellexalGrouping}(cellexalObj, x)
}
\description{
Grouping as cellexalGrouping object will directly check this object.
}
