% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkgrouping.R
\name{checkGrouping,cellexalvrR,character-method}
\alias{checkGrouping,cellexalvrR,character-method}
\title{Grouping as character string will use the cellexalvrR internal grouping for the test.}
\usage{
\S4method{checkGrouping}{cellexalvrR,character}(cellexalObj, x)
}
\description{
Grouping as character string will use the cellexalvrR internal grouping for the test.
}
