% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkgrouping.R
\docType{methods}
\name{checkGrouping}
\alias{checkGrouping}
\alias{checkGrouping,cellexalTime-method}
\title{description of function check}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{x}{an cellexalGrouping or a grouping name.}
}
\description{
checks for NA elements in the table and removes them
}
