% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkgrouping.R
\docType{methods}
\name{checkGrouping}
\alias{checkGrouping}
\alias{checkGrouping,cellexalvrR,character-method}
\alias{checkGrouping,cellexalvrR,cellexalGrouping-method}
\title{cellexalGrouping checks}
\usage{
checkGrouping(cellexalObj, x)

\S4method{checkGrouping}{cellexalvrR,character}(cellexalObj, x)

\S4method{checkGrouping}{cellexalvrR,cellexalGrouping}(cellexalObj, x)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{x}{an cellexalGrouping or a grouping name.}
}
\description{
check the cellexalGrouping
}
\details{
The santity check for a cellexalGrouping object.
}
