% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{checkTime,cellexalTime,ANY-method}
\alias{checkTime,cellexalTime,ANY-method}
\title{The sanity check for a linear grouping.
If problems are detected the problems will be fixed.}
\usage{
\S4method{checkTime}{cellexalTime,ANY}(x, cellexalObj)
}
\description{
The sanity check for a linear grouping.
If problems are detected the problems will be fixed.
}
