% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{checkTime,cellexalTime,cellexalvrR-method}
\alias{checkTime,cellexalTime,cellexalvrR-method}
\title{When presented with a cellexalObj this function makes sure the object fits to the cellexalObj.}
\usage{
\S4method{checkTime}{cellexalTime,cellexalvrR}(x, cellexalObj)
}
\description{
When presented with a cellexalObj this function makes sure the object fits to the cellexalObj.
}
