% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{checkTime}
\alias{checkTime}
\alias{checkTime,cellexalTime-method}
\title{description of function checkTime}
\usage{
\S4method{checkTime}{cellexalTime,ANY}(x, cellexalObj = NULL)

\S4method{checkTime}{cellexalTime,cellexalvrR}(x, cellexalObj = NULL)
}
\arguments{
\item{x}{the cellexalTime object}

\item{cellexalObj}{a cellexalObj fitting to the cellexalTime object}
}
\description{
checks for NA elements in the table and removes them

checks for NA elements in the table and removes them
}
\details{
The sanity check for a linear grouping.
If problems are detected the problems will be fixed.
}
