% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{checkTime}
\alias{checkTime}
\alias{checkTime,cellexalLinear,ANY-method}
\alias{checkTime,cellexalLinear,cellexalvrR-method}
\title{check a cellexalLinear's internals}
\usage{
checkTime(x, cellexalObj = NULL)

\S4method{checkTime}{cellexalLinear,ANY}(x, cellexalObj = NULL)

\S4method{checkTime}{cellexalLinear,cellexalvrR}(x, cellexalObj = NULL)
}
\arguments{
\item{x}{the cellexalLinear object}

\item{cellexalObj}{an optional cellexalvrR object - if given ONLY the drc models are checked.}
}
\description{
checks for NA elements in the checkTime object and removes them
}
\details{
The sanity check for a linear grouping.
If problems are detected the problems will be fixed.
}
