% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVRfiles.R
\docType{methods}
\name{checkVRfiles}
\alias{checkVRfiles}
\alias{checkVRfiles,cellexalvr-method}
\title{description of function checkVRfiles}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{path}{the outpath to check}
}
\description{
checkVRfiles: Checks the existance of all VR specific files and re-runs the export

function if any is missing.
}
