% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrExportFunctions.R
\name{checkVRfiles}
\alias{checkVRfiles}
\title{Makes the base files needed to run the VR environment from a Seurat object}
\usage{
checkVRfiles(cellexalvr, path)
}
\arguments{
\item{path}{the outpath to check}

\item{cellexalObj}{A cellexalvr object}

\item{cellexalObj}{the cellexal object}
}
\description{
Makes the base files needed to run the VR environment from a Seurat object
}
