% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateParts.R
\docType{methods}
\name{cleanParts}
\alias{cleanParts}
\alias{cleanParts,cellexalvrR-method}
\title{cleanParts removes all parts when storing the whole object.}
\arguments{
\item{path}{the outpath}
}
\description{
remove all parts from the file system
}
