% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateParts.R
\docType{methods}
\name{cleanParts}
\alias{cleanParts}
\alias{cleanParts,character-method}
\title{cleanParts removes all parts when storing the whole object.}
\usage{
cleanParts(path)

\S4method{cleanParts}{character}(path)
}
\arguments{
\item{path}{the outpath}
}
\description{
remove all parts from the file system
}
\details{
The parts clean up function is called after a object is saved using the lockedSave function,
as all previousely saved sub parts should have been integrated into this object then.
}
