% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlotHeatmaps.R
\name{clusterGenes,cellexalTime-method}
\alias{clusterGenes,cellexalTime-method}
\title{If the expression matrix is not available the data will be regenreated from the cellexalObj and the cellexalTime object.}
\usage{
\S4method{clusterGenes}{cellexalTime}(
  x,
  deg.genes = NULL,
  info = NULL,
  cellexalObj,
  geneclusters = NULL
)
}
\description{
If the expression matrix is not available the data will be regenreated from the cellexalObj and the cellexalTime object.
}
