% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlotHeatmaps.R
\name{clusterGenes,matrix-method}
\alias{clusterGenes,matrix-method}
\title{The x object will be treated as expression table and clustered.}
\usage{
\S4method{clusterGenes}{matrix}(x, deg.genes = NULL, info = NULL, geneclusters = NULL)
}
\description{
The x object will be treated as expression table and clustered.
}
