% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlotHeatmaps.R
\docType{methods}
\name{clusterGenes}
\alias{clusterGenes}
\alias{clusterGenes,cellexalTime-method}
\title{description of function plot}
\arguments{
\item{x}{either the z-scored matrix or a cellexalTime object}

\item{deg.genes}{a list of genes}

\item{info}{the group to cluster the genes for (list)}

\item{cellexalObj}{if x is a cellexalTime object this is necessary to create the zscored matrix.}
}
\description{
get a - hopefully - optimal grouing of a list of genes
}
\details{
Tries to answer the question: how do these genes differ over the timeline.
It answers this in a graphical, not a statistical way.
Hence you can feed whichever genelist you like into this function.
}
