% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{collapseTime}
\alias{collapseTime}
\alias{collapseTime,cellexalLinear-method}
\title{sum up cells based on the pseudo time}
\usage{
collapseTime(x, to = 1000)

\S4method{collapseTime}{cellexalLinear}(x, to = 1000)
}
\arguments{
\item{x}{the cellexalLinear object}

\item{to}{the max amnount of pseudo time points}
}
\description{
return the color in exactly the right format to color the names
}
\details{
collapse the linear selection to a max of 1000 entries rendering the 
linearSelection incompatible with cellexalvrR
but allowing the correlation of collapsed expression to the pseudo time.
}
