% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{color,cellexalTime-method}
\alias{color,cellexalTime-method}
\title{returns a color vector based on the intern col column and the cell names
Cells that are not part of this selection will get gray(0.6) as color.}
\usage{
\S4method{color}{cellexalTime}(x, names)
}
\description{
returns a color vector based on the intern col column and the cell names
Cells that are not part of this selection will get gray(0.6) as color.
}
