% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{color}
\alias{color}
\alias{color,cellexalvrR-method}
\title{description of function color}
\arguments{
\item{x}{the cellexalTime object}

\item{names}{the cells to color with the time color}
}
\description{
return the color in exactly the right format to color the names
}
\details{
returns a color vector based on the intern col column and the cell names
Not time containing cells will get gray(0.6) as color.
}
