% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{compactTime}
\alias{compactTime}
\alias{compactTime,cellexalTime-method}
\title{description of function plot}
\arguments{
\item{cellexalObj}{if x is a cellexalTime object this is necessary to create the zscored matrix.}
}
\description{
create a global gene z-score and virtual samples over the timeline
}
\details{
creates compaced expression data for all timelines.
}
