% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{compactTimeZscore,cellexalTime,character,cellexalGrouping,cellexalvrR-method}
\alias{compactTimeZscore,cellexalTime,character,cellexalGrouping,cellexalvrR-method}
\title{This function should be run on the main cellexalObj including all data.
It will create (for a subset of genes) a z-scored dataset and will then 
sum up the expression in the time orientation to represent 1000 'cells' in the timeline.}
\usage{
\S4method{compactTimeZscore}{cellexalTime,character,cellexalGrouping,cellexalvrR}(x, deg.genes, info, cellexalObj)
}
\description{
This function should be run on the main cellexalObj including all data.
It will create (for a subset of genes) a z-scored dataset and will then 
sum up the expression in the time orientation to represent 1000 'cells' in the timeline.
}
