% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{compactTimeZscore}
\alias{compactTimeZscore}
\alias{compactTimeZscore,cellexalTime-method}
\title{description of function plot}
\usage{
\S4method{compactTimeZscore}{cellexalTime,character,cellexalGrouping,cellexalvrR}(x, deg.genes, info, cellexalObj)
}
\arguments{
\item{x}{either the z-scored matrix or a cellexalTime object}

\item{deg.genes}{a list of genes  cluster the genes for (list)}

\item{cellexalObj}{if x is a cellexalTime object this is necessary to create the zscored matrix.}
}
\description{
create a global gene z-score and virtual samples over the timeline
}
\details{
This function should be run on the main cellexalObj including all data.
It will create (for a subset of genes) a z-scored dataset and will then 
sum up the expression in the time orientation to represent 1000 'cells' in the timeline.
}
