% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{compareGeneClusters,cellexalTime,cellexalTime,cellexalvrR-method}
\alias{compareGeneClusters,cellexalTime,cellexalTime,cellexalvrR-method}
\title{Time objects are used directly. If the cellexalvrR object has been updated 
in the meantime this function is less saver to use than the character character version.}
\usage{
\S4method{compareGeneClusters}{cellexalTime,cellexalTime,cellexalvrR}(
  x,
  other,
  cellexalObj,
  altGroupNames = NULL,
  color = NULL,
  GOIs = NULL
)
}
\description{
Time objects are used directly. If the cellexalvrR object has been updated 
in the meantime this function is less saver to use than the character character version.
}
