% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{compareGeneClusters,character,character,cellexalvrR-method}
\alias{compareGeneClusters,character,character,cellexalvrR-method}
\title{Time objects are extracted from the cellexalvrR object using the time names instead of the time objects}
\usage{
\S4method{compareGeneClusters}{character,character,cellexalvrR}(
  x,
  other,
  cellexalObj,
  altGroupNames = NULL,
  color = NULL,
  GOIs = NULL
)
}
\description{
Time objects are extracted from the cellexalvrR object using the time names instead of the time objects
}
