% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{compareGeneClusters}
\alias{compareGeneClusters}
\alias{compareGeneClusters,character,character,cellexalvrR-method}
\alias{compareGeneClusters,cellexalTime,cellexalTime,cellexalvrR-method}
\title{compare two cellexalTime objects (interactive mode only)}
\usage{
compareGeneClusters(
  x,
  other,
  cellexalObj,
  altGroupNames = NULL,
  color = NULL,
  GOIs = NULL
)

\S4method{compareGeneClusters}{character,character,cellexalvrR}(
  x,
  other,
  cellexalObj,
  altGroupNames = NULL,
  color = NULL,
  GOIs = NULL
)

\S4method{compareGeneClusters}{cellexalTime,cellexalTime,cellexalvrR}(
  x,
  other,
  cellexalObj,
  altGroupNames = NULL,
  color = NULL,
  GOIs = NULL
)
}
\arguments{
\item{x}{the cellexalTime object}

\item{other}{the other timeline (the parentSelection has to match)}

\item{cellexalObj}{the cellexalObj to add the report to.}

\item{altGroupNames}{you can give alternative group names here}

\item{color}{a vector of colors - one for each group}

\item{GOIs}{a different group of genes - might need to re-run the analysis}
}
\description{
return the color in exactly the right format to color the names
}
\details{
This is a none VR method to compare two different timelines.
The main usage for this function is to run the same genes trough trough one timeline 
that has been split into two sub-sections like young and old or teated and untreated.
This function only allows to compare two cellexalTime objects.
This function highligts which genes follow different expression patterns in both timelines.
}
