% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\docType{methods}
\name{cormat2df}
\alias{cormat2df}
\alias{cormat2df,matrix-method}
\title{convert correlation matrix to data frame (intern)}
\usage{
cormat2df(cors)

\S4method{cormat2df}{matrix}(cors)
}
\arguments{
\item{cors}{A square matrix of pairwise measures}
}
\description{
Converts a pairwise correlation matrix for data.frame
}
\keyword{construction}
\keyword{network}
