% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2D.R
\docType{methods}
\name{correctPath}
\alias{correctPath}
\alias{correctPath,cellexalvrR-method}
\title{description of function correctPath}
\usage{
correctPath(f, cellexalObj)
}
\arguments{
\item{f}{the file path to modify}

\item{cellexalObj}{the cellexal object}
}
\description{
convert the drcPlots2D into rmd format
}
\details{
Correct a file path so that it will be relative to the cellexalvrR@outpath
}
