% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2D.R
\docType{methods}
\name{correctPath}
\alias{correctPath}
\title{convert the drcPlots2D file position to match the log file's position}
\usage{
correctPath(f, cellexalObj)
}
\arguments{
\item{f}{the file path to modify}

\item{cellexalObj}{the cellexal object}
}
\description{
convert the drcPlots2D file position to match the log file's position
}
\details{
Correct a file path so that it will be relative to the cellexalvrR@outpath
}
