% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{createDetailedComparison}
\alias{createDetailedComparison}
\alias{createDetailedComparison,cellexalTime-method}
\title{description of function plot}
\arguments{
\item{x}{the object}

\item{cellexalObj}{the object to get the data from}

\item{deg.genes}{a list of genes to create the report for}

\item{name}{the name for this gene set}

\item{subsets}{for which cell subsets to run the analysis (give me a named list!)}

\item{otherGeneGroupings}{other timeline analyzed genes to in depth descibe the differences}

\item{color}{the colors for the subsets}

\item{geneDescr}{a desription of this gene list - if NULL a list of gene links will be added}
}
\description{
compare a set of genes and compare them in different cells over the same timeline
}
\details{
Tries to answer the question: how do these genes differ over the timeline in two sets of cells.
It answers this in a graphical, not a statistical way.
Hence you can feed whichever genelist you like into this function.

It will cluster these genes based on all data and use this clusering throughout.
Hence the two xy line graphs can directly be compared.
}
