% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{createReport,cellexalTime,cellexalvrR,cellexalGrouping-method}
\alias{createReport,cellexalTime,cellexalvrR,cellexalGrouping-method}
\title{Create the report from a time object and an info grouping object (main function)}
\usage{
\S4method{createReport}{cellexalTime,cellexalvrR,cellexalGrouping}(x, cellexalObj, info, deg.genes = NULL, num.sig = 250)
}
\description{
Create the report from a time object and an info grouping object (main function)
}
