% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{createReport,cellexalTime,cellexalvrR,cellexalTime-method}
\alias{createReport,cellexalTime,cellexalvrR,cellexalTime-method}
\title{Create the report from a time object and an info time object}
\usage{
\S4method{createReport}{cellexalTime,cellexalvrR,cellexalTime}(x, cellexalObj, info, deg.genes = NULL, num.sig = 250)
}
\description{
Create the report from a time object and an info time object
}
