% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{createReport,character,cellexalvrR,character-method}
\alias{createReport,character,cellexalvrR,character-method}
\title{Create the report from a time name and an info name}
\usage{
\S4method{createReport}{character,cellexalvrR,character}(x, cellexalObj, info, deg.genes = NULL, num.sig = 250)
}
\description{
Create the report from a time name and an info name
}
