% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{createReport}
\alias{createReport}
\alias{createReport,cellexalTime-method}
\title{description of function plot}
\usage{
createReport(x, cellexalObj, info, deg.genes = NULL, num.sig = 250)
}
\arguments{
\item{x}{the object}

\item{cellexalObj}{the object to get the data from}

\item{info}{a cellexalvrR grouping info list}

\item{deg.genes}{a list of genes to create the report for}

\item{num.sig}{number of signififcant genes (default 250)}
}
\description{
3D plot the time with a line in the time
}
\details{
This report is the main go to analysis for a timeline.
It will produce gene clusters and plot the mean expression of these genes clusters as beautiful xy plots.
It also populates the CellexalTime geneClusters slot with a list entries containing
the gene clusters (clusters), the max 1000 cells zscored data (matrix) and the location of the main figure (png file path; figure)

After this function two cellexalTime objects can be compared to another using the function compareGeneClusters.
}
