% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{createStats}
\alias{createStats}
\alias{createStats,cellexalTime-method}
\alias{createStats,character,cellexalvrR-method}
\alias{createStats,cellexalTime,cellexalvrR-method}
\title{description of function plot}
\usage{
createStats(x, cellexalObj, num.sig = 250, p.cut = NULL)

\S4method{createStats}{character,cellexalvrR}(x, cellexalObj, num.sig = 250, p.cut = NULL)

\S4method{createStats}{cellexalTime,cellexalvrR}(x, cellexalObj, num.sig = 250, p.cut = NULL)
}
\arguments{
\item{x}{the object}

\item{cellexalObj}{the object to get the data from}

\item{num.sig}{the amountof genes to return as top genes (default 250)}

\item{p.cut}{use a p value cut of instead of a number of significant genes}
}
\description{
calculates linear peason statistics for this timeline and the

expression data in the cellexalvrR object.
}
