% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{createTime}
\alias{createTime}
\alias{createTime,cellexalTime-method}
\title{description of function plot}
\arguments{
\item{x}{the object}

\item{parentSelection}{the selection name this time bases on.}
}
\description{
calculate time based on the internal table a b and c columns
}
