% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineGOIs.R
\docType{methods}
\name{defineGOIs}
\alias{defineGOIs}
\alias{defineGOIs,cellexalvrR-method}
\title{description of function defineGOIs}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{name}{the name of the GIO list (eg TFs or epigenetic)}

\item{genes}{a list of gene symbols that match to the @data rownames}

\item{lables}{a list of lables for the GIO column (default NULL)}

\item{...}{unused}
}
\description{
Allows the user to define (G)enes (O)f (I)nterest lists in the object
}
