% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2D.R
\docType{methods}
\name{drcFiles2HTML}
\alias{drcFiles2HTML}
\title{convert the drcPlots2D into rmd format for the log}
\usage{
drcFiles2HTML(cellexalObj, gInfo, showIDs = TRUE, addOn = NULL)
}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{gInfo}{the return value from cellexalvrR::groupingInfo()}

\item{showIDs}{show the IDs on the plot (default = TRUE)}

\item{addOn}{a text to add in the figure heading (default NULL)}
}
\description{
convert the drcPlots2D into rmd format for the log
}
\details{
drcPlots2D is a function linked to the log functionallity.
This function converts the file paths to Rmd image strings.
}
