% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2Dexpression.R
\docType{methods}
\name{drcFiles2HTMLexpression}
\alias{drcFiles2HTMLexpression}
\title{convert the drcPlots2D into rmd format}
\usage{
drcFiles2HTMLexpression(cellexalObj, drc, GOI, addOn = NULL)
}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{drc}{the drc object to plot the data onto}

\item{GOI}{the return value from cellexalvrR::groupingInfo()}

\item{addOn}{a text to add in the figure heading (default NULL)}
}
\description{
convert the drcPlots2D into rmd format
}
\details{
drcFiles2HTMLexpression is a function linked to the log functionallity.
This function converts the file paths to Rmd image strings.
}
