% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2Dtime.R
\docType{methods}
\name{drcFiles2HTMLtime}
\alias{drcFiles2HTMLtime}
\title{convert the drcPlots2D into rmd format}
\usage{
drcFiles2HTMLtime(cellexalObj, gInfo, showIDs = TRUE, addOn = NULL)
}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{gInfo}{the return value from cellexalvrR::groupingInfo()}

\item{showIDs}{here for compatibility to drcFiles2HTML}

\item{addOn}{a text to add in the figure heading (default NULL)}
}
\description{
convert the drcPlots2D into rmd format
}
\details{
drcFiles2HTMLtime is a function linked to the log functionallity.
This function converts the file paths to Rmd image strings.
}
