% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2Dexpression.R
\docType{methods}
\name{drcPlot2Dexpression}
\alias{drcPlot2Dexpression}
\alias{drcPlot2Dexpression,cellexalvrR,character,character-method}
\title{create one (2D) or three 2D drc plots for the report}
\usage{
drcPlot2Dexpression(cellexalObj, drc = 1, GOI)

\S4method{drcPlot2Dexpression}{cellexalvrR,character,character}(cellexalObj, drc = 1, GOI)
}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{drc}{the edcr (name) to plot on (default the first one)}

\item{GOI}{the genes to plot the mean expression for}
}
\description{
create one (2D) or three 2D drc plots for the report
}
\details{
drcPlot2Dexpression is a function linked to the log functionallity.

This function will create the 2D DRC images for the correlated genes log sections.
}
