% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2D.R
\name{drcPlots2D,cellexalvrR,cellexalGrouping-method}
\alias{drcPlots2D,cellexalvrR,cellexalGrouping-method}
\title{If the gInfo is a cellexalGrouping object this object will be used directly}
\usage{
\S4method{drcPlots2D}{cellexalvrR,cellexalGrouping}(cellexalObj, gInfo, GOIs = NULL, showIDs = TRUE)
}
\description{
If the gInfo is a cellexalGrouping object this object will be used directly
}
