% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2D.R
\name{drcPlots2D,cellexalvrR,character-method}
\alias{drcPlots2D,cellexalvrR,character-method}
\title{The cellexalvrR internal grouping will be used if gInfo is a string}
\usage{
\S4method{drcPlots2D}{cellexalvrR,character}(cellexalObj, gInfo, GOIs = NULL, showIDs = TRUE)
}
\description{
The cellexalvrR internal grouping will be used if gInfo is a string
}
