% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2D.R
\docType{methods}
\name{drcPlots2D}
\alias{drcPlots2D}
\alias{drcPlots2D,cellexalvrR,character-method}
\alias{drcPlots2D,cellexalvrR,cellexalGrouping-method}
\title{create one (2D) or three 2D drc plots for the report}
\usage{
drcPlots2D(cellexalObj, gInfo, showIDs = TRUE)

\S4method{drcPlots2D}{cellexalvrR,character}(cellexalObj, gInfo, showIDs = TRUE)

\S4method{drcPlots2D}{cellexalvrR,cellexalGrouping}(cellexalObj, gInfo, showIDs = TRUE)
}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{gInfo}{the grouping name or the cellexalGrouping object}

\item{showIDs}{plot the group IDs on the figure (default= TRUE)}
}
\description{
create one (2D) or three 2D drc plots for the report
}
\details{
drcPlots2D is a function linked to the log functionallity.
This function will create the 2D DRC images for the log sections.
}
