% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2Dlinear.R
\docType{methods}
\name{drcPlots2Dlinear}
\alias{drcPlots2Dlinear}
\alias{drcPlots2Dlinear,cellexalvrR,cellexalLinear-method}
\alias{drcPlots2Dlinear,cellexalvrR,cellexalGrouping-method}
\title{create two 2D drc plots for the report}
\usage{
drcPlots2Dlinear(cellexalObj, gInfo)

\S4method{drcPlots2Dlinear}{cellexalvrR,cellexalLinear}(cellexalObj, gInfo)

\S4method{drcPlots2Dlinear}{cellexalvrR,cellexalGrouping}(cellexalObj, gInfo)
}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{gInfo}{either a cellexalLinear or a cellexalGrouping object}
}
\description{
create two 2D drc plots for the report
}
\details{
drcPlots2Dlinear is a function linked to the log functionallity.

This function will create the 2D DRC images for the logLinear selection sections.
}
