% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2Dtime.R
\docType{methods}
\name{drcPlots2Dtime}
\alias{drcPlots2Dtime}
\alias{drcPlots2Dtime,cellexalvrR-method}
\title{description of function drcPlot2D}
\usage{
drcPlots2Dtime(cellexalObj, gInfo)
}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{gInfo}{the return value from cellexalvrR::groupingInfo()}
}
\description{
create two 2D drc plots for the report
}
\details{
drcPlots2Dtime is a function linked to the log functionallity.

This function will create the 2D DRC images for the logTimeline sections.
}
