% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrop.R
\docType{methods}
\name{entrop}
\alias{entrop}
\alias{entrop,cellexalvrR-method}
\title{description of function entrop}
\usage{
entrop(d)
}
\arguments{
\item{d}{a vector of group names ( class character)}
}
\description{
description of function entrop
}
