% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_data_files.R, R/data.R
\docType{data}
\name{Epigenetic}
\alias{Epigenetic}
\alias{Epigenetic}
\title{A simple table containing the data from http://epifactors.autosome.ru/ as from 21st September 2017}
\format{data.frame}
\usage{
Epigenetic

Epigenetic
}
\description{
This table can be used to create the epigenetics MDS objects.

This table can be used to create the epigenetics MDS objects.
}
\keyword{datasets}
