% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportFunctions.R
\name{export2cellexalvr}
\alias{export2cellexalvr}
\title{Creates the base files needed to run the VR environment}
\usage{
export2cellexalvr(cellexalObj, path, forceDB = F, VRpath = NULL)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{forceDB}{re-write the db even if it exisis (default =F)}
}
\description{
Creates the base files needed to run the VR environment
}
