% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_GOIs.R
\docType{methods}
\name{export_GOIs}
\alias{export_GOIs}
\alias{export_GOIs,cellexalvrR-method}
\title{description of function export_GOIs}
\arguments{
\item{x}{the cellexal object}

\item{GOIs}{the genes of interest}

\item{grouping}{the grouping the hetmap should be using (a cellexalVR selection file will be produced)}

\item{path}{the outpath for the files}

\item{colorF}{a function returning the colors in the order they should appear on the heatmap (default = rainbow)}
}
\description{
export a list of genes of interest to be loaded as heatmaps in cellexalVR
}
\details{
Normally a scientis has a number of genes that are of high interest for the spüecific data set.
This function can be used to export this gene list as a buch of hetmap input files for cellexalVR.

These files can be loaded from the cellexalVR console by using the command
GOIs path
}
