% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_UserGrouping.R
\docType{methods}
\name{export_UserGrouping}
\alias{export_UserGrouping}
\alias{export_UserGrouping,cellexalvrR-method}
\title{description of function export_UserGrouping}
\arguments{
\item{x}{the cellexal object}

\item{colname}{the userGrouping column name}

\item{outfile}{the out file}

\item{colorF}{a function returning the colors in the order they should appear on the heatmap (default = rainbow)}
}
\description{
export a list of genes of interest to be loaded as heatmaps in cellexalVR
}
