% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{exportSelection}
\alias{exportSelection}
\alias{exportSelection,cellexalTime-method}
\title{description of function exportSelection}
\usage{
\S4method{exportSelection}{cellexalTime}(x, fname)
}
\arguments{
\item{x}{the object}

\item{fname}{the file to write the info to}
}
\description{
Export this time as selection file
}
\details{
This function exports the changes selection file including the linear clusters
so that the information can be read from the VR process.
}
