% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUserGroups4vr.R
\docType{methods}
\name{exportUserGroups4vr}
\alias{exportUserGroups4vr}
\alias{exportUserGroups4vr,cellexalvrR-method}
\title{description of function exportUserGroups4vr}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{path}{the outpath}
}
\description{
Creates a summary file for the vr process Creates a file groupings_info.txt in the

outfolder that contains the group name (in the R object) the numer of groups in the

selection and the number of cells in the whole group.
}

