% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUserGroups4vr.R
\docType{methods}
\name{exportUserGroups4vr,character-method}
\alias{exportUserGroups4vr,character-method}
\title{load object and run exportUserGroups4vr}
\usage{
\S4method{exportUserGroups4vr}{character}(cellexalObj, path)
}
\arguments{
\item{cellexalObj}{the cellexalObj.RData file}

\item{path}{the outpath}
}
\description{
simply preload the cellexalObj before running the cellexalvrR specififc function
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: cellexalvrR
}}

