% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptHelpers.R
\name{file2Script}
\alias{file2Script}
\title{Convert a file 'A/B/C.txt' into the string 'file.path('A','B','C.txt')'.}
\usage{
file2Script(path, mustWork = FALSE)
}
\arguments{
\item{path}{the file to convert}

\item{mustWork}{default = FALSE}
}
\description{
Convert a file 'A/B/C.txt' into the string 'file.path('A','B','C.txt')'.
}
