% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\docType{methods}
\name{filename}
\alias{filename}
\alias{filename,character-method}
\title{create a filename without spaces from a list of strings}
\usage{
filename(str)

\S4method{filename}{character}(str)
}
\arguments{
\item{str}{a string vector}
}
\description{
simple function to join by '.' and remove whitespace.
}
\details{
Filename simple pastes a vector of characters using '.' and removes all whitespace.

Simple function to unclutter the R code.
}
