% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceAbsoluteUniqueSample.R
\docType{methods}
\name{forceAbsoluteUniqueSample}
\alias{forceAbsoluteUniqueSample}
\alias{forceAbsoluteUniqueSample,cellexalvr-method}
\title{description of function forceAbsoluteUniqueSample}
\arguments{
\item{x}{the vector of sample names}

\item{separator}{TEXT MISSING default='_'}
}
\description{
A simple function that forces all samplenames to be unique.
}
