% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cellexalvrR.R
\name{forceAbsoluteUniqueSample}
\alias{forceAbsoluteUniqueSample}
\title{This function adds _<id> to all duplicate values thereby enforcing uniques.}
\usage{
forceAbsoluteUniqueSample(x, separator = "_")
}
\description{
This function adds _<id> to all duplicate values thereby enforcing uniques.
}
