% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\docType{methods}
\name{get.genes.cor.to}
\alias{get.genes.cor.to}
\alias{get.genes.cor.to,cellexalvrR-method}
\title{description of function get.genes.cor.to}
\usage{
\S4method{get.genes.cor.to}{cellexalvrR}(cellexalObj, gname,
  output = NULL, is.smarker = F, cpp = T)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{gname}{The required gene}

\item{output}{the outfile}

\item{is.smarker}{Whether the supplied gene is a surface marker (default =F)}

\item{cpp}{use the c++ cor implementation (default = TRUE)}
}
\description{
Gets positively and negatively correlated genes to a chosen gene
}
\examples{
print (get.genes.cor.to ( cellexalObj, 'Gata1'))
}
\keyword{correlation}
