% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentials.R
\docType{methods}
\name{getDifferentials}
\alias{getDifferentials}
\alias{getDifferentials,cellexalvrR-method}
\title{description of function getDifferentials}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{deg.method}{The method to use to find DEGs ( 'anova', 'edgeR', 'MAST' or 'Seurat')}

\item{num.sig}{number of differnetial genes to return (250)}

\item{Log}{log the results (default=TRUE)}

\item{logfc.threshold}{the Seurat logfc.threshold option (default here 1 vs 0.25 in Seurat)}
}
\description{
Creates a heatmap from a selection of groups
The Seurat based statsictsics is applied only to genes expressed in at least 1 percent of the cells.
}
\keyword{DEGs}
