% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentials.R
\name{getDifferentials}
\alias{getDifferentials}
\title{Creates a heatmap from a selection of groups}
\usage{
getDifferentials(cellexalObj, cellidfile, deg.method = c("anova", "edgeR",
  "MAST"), num.sig)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{deg.method}{The method to use to find DEGs}

\item{numsig}{The number of differentials to be returned}
}
\description{
Creates a heatmap from a selection of groups
}
\keyword{DEGs}
