% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinearSelection.R
\docType{methods}
\name{getLinearSelection}
\alias{getLinearSelection}
\alias{getLinearSelection,cellexalvrR,character-method}
\alias{getLinearSelection,cellexalvrR,cellexalLinear-method}
\title{get an existing linearSelection or throw an error}
\usage{
getLinearSelection(cellexalObj, name)

\S4method{getLinearSelection}{cellexalvrR,character}(cellexalObj, name)

\S4method{getLinearSelection}{cellexalvrR,cellexalLinear}(cellexalObj, name)
}
\arguments{
\item{cellexalObj}{the object to get the data from}

\item{name}{the name for this linearSelection}
}
\description{
Accessor function to get a cellexalLinear object from a cellexalvrR object by name.
}
