% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentials.R
\docType{methods}
\name{getStatsMethod}
\alias{getStatsMethod}
\alias{getStatsMethod,cellexalvrR-method}
\title{get the method to calculate differential gene expression}
\usage{
getStatsMethod(x, method = "Seurat")
}
\arguments{
\item{x}{the cellexalvrR object}
}
\description{
get the default stat method rpreviousely set with setStatsMethod()
}
