% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTime.R
\docType{methods}
\name{getTime}
\alias{getTime}
\alias{getTime,cellexalTime-method}
\title{get a timeline or throw an error}
\usage{
\S4method{getTime}{cellexalvrR,character}(cellexalObj, name)

\S4method{getTime}{cellexalvrR,cellexalTime}(cellexalObj, name)
}
\arguments{
\item{cellexalObj}{the object to get the data from}

\item{name}{the name for this timeline}
}
\description{
compare a set of genes and compare them in different cells over the same timeline

compare a set of genes and compare them in different cells over the same timeline
}
