% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTime.R
\docType{methods}
\name{getTime}
\alias{getTime}
\alias{getTime,cellexalvrR,character-method}
\alias{getTime,cellexalvrR,cellexalTime-method}
\title{get an existing timeline or throw an error}
\usage{
getTime(cellexalObj, name)

\S4method{getTime}{cellexalvrR,character}(cellexalObj, name)

\S4method{getTime}{cellexalvrR,cellexalTime}(cellexalObj, name)
}
\arguments{
\item{cellexalObj}{the object to get the data from}

\item{name}{the name for this timeline}
}
\description{
Accessor function to get a cellexalTime object from a cellexalvrR object by name.
}
