% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group2HTMLtable.R
\docType{methods}
\name{group2HTMLtable}
\alias{group2HTMLtable}
\alias{group2HTMLtable,cellexalvrR-method}
\title{create a log table from a grouping}
\usage{
\S4method{group2HTMLtable}{cellexalvrR,list}(x, info, names = NULL)
}
\arguments{
\item{x, }{cellexalvrR object}

\item{info}{the grouping info to convert}

\item{names}{optional vector of names for the groupings (grouping info is numeric)}
}
\description{
create an HTML table from a user grouping
}
\details{
Log related function - convert a session into a html table that can be added to the log
}
