% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupingInfo.R
\docType{methods}
\name{groupingInfo}
\alias{groupingInfo}
\alias{groupingInfo,cellexalvrR-method}
\title{description of function groupingInfo}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{gname}{The optional group name to get info on a specific grouping (not the last)}

\item{cellexalObj, }{TEXT MISSING}

\item{gname}{TEXT MISSING default=NULL}
}
\description{
returns the information stored for the last grouping read
}
\keyword{groupingInfo}

