% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupingInfo.R
\name{groupingInfo}
\alias{groupingInfo}
\title{returns the information stored for the last grouping read}
\usage{
groupingInfo(cellexalObj, gname = NULL)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{gname}{The optional group name to get info on a specific grouping (not the last)}
}
\description{
returns the information stored for the last grouping read
}
\keyword{groupingInfo}
