% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cellexalvrR.R
\name{H5Anno2df,H5File-method}
\alias{H5Anno2df,H5File-method}
\title{Convert e.g. a h5ad obs object to a R::data.frame.}
\usage{
\S4method{H5Anno2df}{H5File}(x, slotName, namecol = NULL, onlyStrings = FALSE)
}
\description{
Convert e.g. a h5ad obs object to a R::data.frame.
}
