% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cellexalvrR.R, R/processVelocytoPyResults.R
\docType{methods}
\name{H5Anno2df}
\alias{H5Anno2df}
\alias{H5Anno2df,cellexalvrR-method}
\title{description of function H5Anno2df}
\usage{
H5Anno2df(x, slotName, namecol = NULL, onlyStrings = FALSE)

H5Anno2df(x, slotName, namecol = NULL, onlyStrings = FALSE)
}
\arguments{
\item{x}{the H5 object}

\item{slotName}{the H5 entity tro convert to a data.frame}

\item{namecol}{the (optional) rownames column for the data}

\item{onlyStrings}{return only columns that not only contain numbers (default FALSE)}
}
\description{
convert a H5 annotation (any name) table to a data table

convert a H5 annotation (any name) table to a data table
}
\details{
Here two Velocyto file results (actually andata hdf5 files) 
one containing the var gene analysis with a meaningful clustering
and one with all genes. This function will combine the drc models from the var gene analysis 
and it's clustering with the gene expressions in the bigger file.

The output will be a cellexalvrR object.
}
