% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cellexalvrR.R
\name{H5Anno2df}
\alias{H5Anno2df}
\title{convert a H5 annotation (any name) table to a data table}
\usage{
H5Anno2df(x, slotName, namecol = NULL, onlyStrings = FALSE)
}
\arguments{
\item{x}{the H5 object}

\item{slotName}{the H5 entity tro convert to a data.frame}

\item{namecol}{the (optional) rownames column for the data}

\item{onlyStrings}{return only columns that not only contain numbers (default FALSE)}
}
\description{
convert a H5 annotation (any name) table to a data table
}
