% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03cellexalGrouping.R
\name{HTMLtable,cellexalGrouping-method}
\alias{HTMLtable,cellexalGrouping-method}
\title{Create an HTML table from a cellexalGrouping.
These tables are part of the log html's and show group color as well as cell count, VR- and R-ID.}
\usage{
\S4method{HTMLtable}{cellexalGrouping}(x)
}
\description{
Create an HTML table from a cellexalGrouping.
These tables are part of the log html's and show group color as well as cell count, VR- and R-ID.
}
