% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{HTMLtable}
\alias{HTMLtable}
\alias{HTMLtable,cellexalTime-method}
\title{description of function color}
\arguments{
\item{x}{the cellexalTime object}
}
\description{
return the HTMLtable in exactly the right format
}
\details{
returns a color vector based on the intern col column and the cell names
Not time containing cells will get gray(0.6) as color.
}
