% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateParts.R
\docType{methods}
\name{integrateParts}
\alias{integrateParts}
\alias{integrateParts,cellexalvrR-method}
\title{integrate the parts that have been split from the main object}
\usage{
integrateParts(x, path = NULL)

\S4method{integrateParts}{cellexalvrR}(x, path = NULL)
}
\arguments{
\item{x}{the cellexalObj to add to}

\item{path}{the path where the parts are stored (normally the @outpath)}
}
\description{
CellexalvrR object can be saved partially. This function loads them.
This function is called by the loadObject function. Nothing the user has to think about!
}
\details{
The function integrateParts is one main part in the load and save speed up implemented for cellexalvrR.

Instead of saving all data in each VR script call only the modified data is saved.
This function loads all separately saved parts into the main cellexalvrR object.
}
