% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{invert}
\alias{invert}
\alias{invert,cellexalTime-method}
\title{reverse the time order for a cellexalTime object}
\usage{
invert(x)

\S4method{invert}{cellexalTime}(x)
}
\arguments{
\item{x}{the object}
}
\description{
reverse the time orientaion
}
\details{
inverts the time in the object.

Occational the algorithm fails and inverts the time.
This function is used to easily invert a timeline that 
runs from end to start instead the otherway around.
}
