% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadObject.R
\docType{methods}
\name{loadObject}
\alias{loadObject}
\alias{loadObject,cellexalvr-method}
\title{description of function loadObject}
\arguments{
\item{fname}{the file to load}

\item{maxwait}{TEXT MISSING default=50}
}
\description{
loads a cellexalvr object fron a file, but waits on 
all systems before another process has finished writing the file
unsing the lockedSave() function.
}
