% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockedSave.R
\docType{methods}
\name{loadObject,cellexalvrR-method}
\alias{loadObject,cellexalvrR-method}
\title{dummy function just returning the cellexalvrR object.}
\usage{
\S4method{loadObject}{cellexalvrR}(fname, maxwait = 50)
}
\arguments{
\item{fname}{the file to load or a cellexalvr object}

\item{maxwait}{stop after maxwait seconds default=50}
}
\description{
just returns the cellexalObj
}
\section{Methods (by class)}{
\itemize{
\item \code{cellexalvrR}: cellexalvrR
}}

\keyword{load}
