% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockedSave.R
\docType{methods}
\name{lockedLoad}
\alias{lockedLoad}
\alias{lockedLoad,character-method}
\title{description of function lockedSave}
\arguments{
\item{cellexalObj}{the file containing the cellexalObj data}
}
\description{
Loading the RData in the VR tool might create a problem. Hence this function will
save the cellexalObj in a controlled way.
}
\keyword{lockedSave}
