% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockedSave.R
\name{lockedSave}
\alias{lockedSave}
\title{Saving the RData in the VR tool might create a problem. Hence this function will save the cellexalObj in a controlled way.}
\usage{
lockedSave(cellexalObj, path)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{path}{the output path}
}
\description{
Saving the RData in the VR tool might create a problem. Hence this function will save the cellexalObj in a controlled way.
}
\keyword{lockedSave}
