% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFigure.R
\docType{methods}
\name{logFigure}
\alias{logFigure}
\alias{logFigure,cellexalvrR-method}
\title{description of function logFigure}
\usage{
\S4method{logFigure}{cellexalvrR}(cellexalObj, png, text = NULL, ...)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{png}{the VR generated network (png)}

\item{text}{add a simmple text (default NULL)}

\item{...}{options you want to send to the ontologyLogPage() function #function definition in file 'ontologyLogPage.R'}
}
\description{
Log a figure from VR. This function is used to store new screenshots from VR.
}
\details{
logFigure is a VR helper funtion that stores one figure in the log document.
}
