% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFigure.R
\docType{methods}
\name{logFigure}
\alias{logFigure}
\alias{logFigure,cellexalvrR-method}
\title{add one figure file to the log structure}
\usage{
logFigure(cellexalObj, png, text = NULL)

\S4method{logFigure}{cellexalvrR}(cellexalObj, png, text = NULL)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{png}{any png figure file}

\item{text}{add a simmple text (default NULL)}
}
\description{
Log a figure from VR. This function is used to store new screenshots from VR.
}
\details{
logFigure is a VR helper funtion that stores one figure in the log document.
}
