% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logHeatmap.R
\docType{methods}
\name{logHeatmap}
\alias{logHeatmap}
\alias{logHeatmap,cellexalvrR-method}
\title{description of function logHeatmap}
\usage{
logHeatmap(cellexalObj, genes, png, grouping)
}
\arguments{
\item{cellexalObj}{teh cellexalvrR object}

\item{genes}{the genes displayed on the heatmap}

\item{png}{the VR generated heatmap (png)}

\item{grouping}{the grouping file used to create this heatmap}
}
\description{
create one Heatmap session report page
}
