% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logHeatmap.R
\docType{methods}
\name{logHeatmap}
\alias{logHeatmap}
\alias{logHeatmap,cellexalvrR-method}
\alias{logHeatmap,character-method}
\title{add the VR created heatmap to the log system}
\usage{
logHeatmap(cellexalObj, genes = NULL, png, grouping, ...)

\S4method{logHeatmap}{cellexalvrR}(cellexalObj, genes = NULL, png, grouping, ...)

\S4method{logHeatmap}{character}(cellexalObj, genes = NULL, png, grouping, ...)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{genes}{the genes displayed on the heatmap (default NULL)}

\item{png}{the VR generated heatmap (png)}

\item{grouping}{the grouping file used to create this heatmap}

\item{...}{options you want to send to the ontologyLogPage() function}
}
\description{
Add a VR created heatmap into the log.
}
\details{
logHeatmap will create a section in the log document including 

(1) the DRC the grouping was selected from (colored 2D)

(2) the heatmap itself

(3) a GO analysis of the genes displayed in the heatmap (using ontologyLogPage())
}
