% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNetwork.R
\docType{methods}
\name{logNetwork}
\alias{logNetwork}
\alias{logNetwork,cellexalvrR-method}
\title{description of function logNetwork}
\usage{
\S4method{logNetwork}{cellexalvrR}(cellexalObj, genes = NULL, png,
  grouping, ...)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{genes}{the genes displayed on the network}

\item{png}{the VR generated network (png)}

\item{grouping}{the grouping file used to create this network}

\item{...}{options you want to send to the ontologyLogPage() function #function definition in file 'ontologyLogPage.R'}
}
\description{
create one Network page in the session report
}
\details{
logNetwork is a VR helper funtion that stores one network in the log document.
}
