% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNetwork.R
\docType{methods}
\name{logNetwork}
\alias{logNetwork}
\alias{logNetwork,cellexalvrR-method}
\alias{logNetwork,character-method}
\title{add one network to the cellexalvrR log system}
\usage{
logNetwork(cellexalObj, genes = NULL, grouping, ...)

\S4method{logNetwork}{cellexalvrR}(cellexalObj, genes = NULL, grouping, ...)

\S4method{logNetwork}{character}(cellexalObj, genes = NULL, grouping, ...)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{genes}{the genes displayed on the network}

\item{grouping}{the grouping file used to create this network}

\item{...}{options you want to send to the ontologyLogPage() function}

\item{the}{VR generated network ()}
}
\description{
create one Network page in the session report
}
\details{
logNetwork is a VR helper funtion that stores one network plot into the log document.
}
