% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNetworkPNG.R
\docType{methods}
\name{logNetworkPNG}
\alias{logNetworkPNG}
\alias{logNetworkPNG,cellexalvrR-method}
\alias{logNetworkPNG,character-method}
\title{add one network to the cellexalvrR log system}
\usage{
logNetworkPNG(cellexalObj, genes = NULL, png, grouping, ...)

\S4method{logNetworkPNG}{cellexalvrR}(cellexalObj, genes = NULL, png, grouping, ...)

\S4method{logNetworkPNG}{character}(cellexalObj, genes = NULL, png, grouping, ...)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{genes}{the genes displayed on the network}

\item{png}{the VR generated network (png)}

\item{grouping}{the grouping file used to create this network}

\item{...}{options you want to send to the ontologyLogPage() function}
}
\description{
create one Network page in the session report
}
\details{
logNetworkPNG is a VR helper funtion that stores one network plot into the log document.
}
