% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNetworkStats.R
\docType{methods}
\name{logNetworkStats}
\alias{logNetworkStats}
\alias{logNetworkStats,cellexalvrR-method}
\alias{logNetworkStats,character-method}
\title{add one network to the cellexalvrR log system}
\usage{
logNetworkStats(cellexalObj, genes = NULL, grouping, ...)

\S4method{logNetworkStats}{cellexalvrR}(cellexalObj, genes = NULL, grouping, ...)

\S4method{logNetworkStats}{character}(cellexalObj, genes = NULL, grouping, ...)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{genes}{the genes displayed on the network}

\item{grouping}{the grouping file used to create this network}

\item{...}{options you want to send to the ontologyLogPage() function}

\item{the}{VR generated network ()}
}
\description{
create one Network page in the session report
}
\details{
logNetworkStats is a VR helper funtion that stores one network plot into the log document.
}
