% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logStatResult.R
\docType{methods}
\name{logStatResult}
\alias{logStatResult}
\alias{logStatResult,cellexalvrR-method}
\title{add a table into the session log}
\usage{
logStatResult(x, method, data, col = NULL)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{method}{the stats method used (no spaces please)}

\item{data}{the results table}

\item{col}{the p value column to plot the -log10 histogram}
}
\description{
add small section in the log file including a stats data download link (coma separted table file)
}
