% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logStatResult.R
\docType{methods}
\name{logStatResult}
\alias{logStatResult}
\alias{logStatResult,cellexalvrR-method}
\title{make the stats table available in the log system}
\usage{
logStatResult(x, method, data, col = NULL)

\S4method{logStatResult}{cellexalvrR}(x, method, data, col = NULL)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{method}{the stats method used (no spaces please)}

\item{data}{the results table}

\item{col}{the p value column to plot the -log10 histogram}
}
\description{
Create a log file part including a stats data download link (coma separted table file)
}
\details{
logStatResult will create a clickable link in the log document 
where the user can download the stats table from.
}
