% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTimeLine.R
\name{logTimeLine,character-method}
\alias{logTimeLine,character-method}
\title{Load the cellexalObj from a file instead of using it as R object.}
\usage{
\S4method{logTimeLine}{character}(
  cellexalObj,
  stats,
  genes = NULL,
  info,
  png,
  timeInfo,
  GOIs = NULL,
  text = NULL
)
}
\description{
Load the cellexalObj from a file instead of using it as R object.
}
