% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTimeLine.R
\docType{methods}
\name{logTimeLine,character-method}
\alias{logTimeLine,character-method}
\title{description of function logTimeLine}
\usage{
\S4method{logTimeLine}{character}(cellexalObj, stats, genes, info)
}
\arguments{
\item{cellexalObj}{the cellexalvrR file}

\item{stats}{the correlation statistics}

\item{genes}{the genes to display on the heatmap}

\item{info}{the grouping information list}
}
\description{
create one Heatmap session report page
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: cellexalvrR
}}

