% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{make.cell.meta.from.df}
\alias{make.cell.meta.from.df}
\title{Creates a meta cell matrix from a supplied dataframe from required fields}
\usage{
make.cell.meta.from.df(metad, rq.fields)
}
\arguments{
\item{metad}{A dataframe of per cell metadata}

\item{rq.fields}{A vector of name specifiying which columns should me made into metadata}
}
\description{
Creates a meta cell matrix from a supplied dataframe from required fields
}
\keyword{cell}
\keyword{metadata}
