% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringFunctions.R
\docType{methods}
\name{make.cellexalvr.heatmap}
\alias{make.cellexalvr.heatmap}
\alias{make.cellexalvr.heatmap,cellexalvrR-method}
\title{description of function make.cellexalvr.heatmap}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{numsig}{The number of differentials to be returned}

\item{outfile}{The name of the output file}

\item{num.sig}{TEXT MISSING}
}
\description{
Creates a heatmap from a selection of groups
}
\keyword{heatmap}

