% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cellexalvr.heatmap.list.R
\docType{methods}
\name{make.cellexalvr.heatmap.list}
\alias{make.cellexalvr.heatmap.list}
\alias{make.cellexalvr.heatmap.list,cellexalvrR-method}
\title{Create the gene list files for VR}
\arguments{
\item{cvrObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs or the grouping name}

\item{num.sig}{The number of differentials to be returned}

\item{outfile}{The name of the output file}

\item{stats_method}{the stats method to use see getDifferentials()}
}
\description{
Creates an outfile for CellexalVR containing a list of genes from which a heatmap is made in VR
Internally this function uses the getDifferentials() function and just writes the output of that into the outfile.
}
\keyword{differential,}
\keyword{gene}
\keyword{list}
