% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\name{make.cellexalvr.heatmap.list}
\alias{make.cellexalvr.heatmap.list}
\title{Creates a list of genes from which a heatmap is made in Unity}
\usage{
make.cellexalvr.heatmap.list(cvrObj, cellidfile, num.sig, outfile)
}
\arguments{
\item{cellidfile}{file containing cell IDs}

\item{outfile}{The name of the output file}

\item{cellexalObj}{A cellexalvr object}

\item{numsig}{The number of differentials to be returned}
}
\description{
Creates a list of genes from which a heatmap is made in Unity
}
\keyword{heatmap}
