% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\name{make.cellexalvr.heatmap}
\alias{make.cellexalvr.heatmap}
\title{Creates a heatmap from a selection of groups}
\usage{
make.cellexalvr.heatmap(cvrObj, cellidfile, num.sig, outfile)
}
\arguments{
\item{cellidfile}{file containing cell IDs}

\item{outfile}{The name of the output file}

\item{cellexalObj}{A cellexalvr object}

\item{numsig}{The number of differentials to be returned}
}
\description{
Creates a heatmap from a selection of groups
}
\keyword{heatmap}
