% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\name{make.cellexalvr.network}
\alias{make.cellexalvr.network}
\title{Creates a network from selected groups for selected genes}
\usage{
make.cellexalvr.network(cellexalObj, cellidfile, outpath, cutoff.ggm = 0.8)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{cutoff.ggm}{The cutoff for the correlation (default = 0.8)}

\item{outfile}{The name of the output file}
}
\description{
Creates a network from selected groups for selected genes
}
\keyword{construction}
\keyword{network}
