% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCellexalVRObj.R
\name{MakeCellexalVRObj,dgCMatrix-method}
\alias{MakeCellexalVRObj,dgCMatrix-method}
\title{MakeCellexalVRObj creates a CellexalVR object from a sparse matrix object.}
\usage{
\S4method{MakeCellexalVRObj}{dgCMatrix}(
  exdata,
  drc.list,
  specie = c("mouse", "human"),
  cell.metadata = NULL,
  facs.data = NULL
)
}
\description{
MakeCellexalVRObj creates a CellexalVR object from a sparse matrix object.
}
