% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCellexalVRObj.R
\docType{methods}
\name{MakeCellexalVRObj}
\alias{MakeCellexalVRObj}
\alias{MakeCellexalVRObj,matrix-method}
\title{Create a cellexalvrR object from a matrix object.}
\usage{
MakeCellexalVRObj(
  exdata,
  drc.list,
  specie = c("mouse", "human"),
  cell.metadata = NULL,
  facs.data = NULL
)
}
\arguments{
\item{exdata}{A matrix of expression values (required). Colnames are cell IDs, rownames are unique gene names}

\item{drc.list}{A list of 3-column DRC coords (at least one required)}

\item{specie}{Specify whether data is from mouse or human (required)}

\item{cell.metadata}{Required meta data for cells}

\item{facs.data}{Surface marker intensities from index sorted cells}
}
\description{
Creates a cellexalvr objext with the data given
}
\keyword{cellexalvrR}
\keyword{create}
\keyword{object}
