% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCellexalVRObj.R
\name{MakeCellexaVRObj}
\alias{MakeCellexaVRObj}
\title{Creates a cellexalvr objext with the data given}
\usage{
MakeCellexaVRObj(exdata, mds.list, specie = c("mouse", "human"),
  cell.metadata = NULL, facs.data = matrix())
}
\arguments{
\item{exdata}{A matrix of expression values (required). Colnames are cell IDs, rownames are unique gene names}

\item{mds.list}{A list of 3-column MDS coords (at least one required)}

\item{specie}{Specify whether data is from mouse of human (required)}

\item{cell.metadata}{Required meta data for cells}

\item{facs}{Surface marker intensities from index sorted cells}
}
\description{
Creates a cellexalvr objext with the data given
}
\keyword{heatmap}
