% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_gene_links.R
\docType{methods}
\name{md_gene_links}
\alias{md_gene_links}
\alias{md_gene_links,cellexalvrR-method}
\title{Create a binned annotation column from numeric data}
\usage{
md_gene_links(genes)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{group}{the group (defaule = 'nUMI'}

\item{where}{in the samples (sample) or annotation (gene) data frame (sample)}

\item{colFun}{colory function default=  gplots::bluered}
}
\description{
Bin the UMI data into 13 bins for plotting and define a blue <- red color gradient
}
\details{
convert a list of genes into a collapsed markdown text that the user can expand.
}
