% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsPlot2D.R
\docType{methods}
\name{mdsPlots2D}
\alias{mdsPlots2D}
\alias{mdsPlots2D,cellexalvrR-method}
\title{description of function mdsPlot2D}
\usage{
mdsPlots2D(cellexalObj, gInfo)
}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{gInfo}{the return value from cellexalvrR::groupingInfo()}
}
\description{
create two 2D mds plots for the report
}
