% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNextCoolHeatmapFeature.R
\docType{methods}
\name{myNextCoolHeatmapFeature}
\alias{myNextCoolHeatmapFeature}
\alias{myNextCoolHeatmapFeature,cellexalvrR,character,character-method}
\title{empty function to be changed by any user (future stuff)}
\usage{
myNextCoolHeatmapFeature(x, selection, outfile)

\S4method{myNextCoolHeatmapFeature}{cellexalvrR,character,character}(x, selection, outfile)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{selection}{the selection file linked to the heatmap}

\item{outfile}{the outfile the VR process is waiting for}
}
\description{
Unused function for future implementations.
}
\details{
This function is linked to a button on the VR heatmaps.
In the default cellexalvrR version this function is a blank to allow non VR Bioinformaticians
to start a function they miss and want to implement themselves.

Unimplemented in VR. Future stuff.
}
