% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveEntropy.R
\docType{methods}
\name{NaiveEntropy}
\alias{NaiveEntropy}
\alias{NaiveEntropy,cellexalvrR-method}
\title{description of function NaiveEntropy}
\usage{
NaiveEntropy(x, gvect, drc, n = 20, sumFunc = sum())
}
\arguments{
\item{x}{the cellexalvrR object}

\item{gvect}{a vector with group ids}

\item{drc}{the dimension reduction name to process}

\item{n}{the amount of nearest neighbors (default= 20)}

\item{sumFunc}{function to sum up the single entropy values (default sum)}
}
\description{
Nearest neighbors based entropy calculation.
}
\details{
Calculates a simple entropy value for a grouping vector and a 3D drc data.
}
