% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveEntropySpherical.R
\docType{methods}
\name{NaiveEntropySpherical}
\alias{NaiveEntropySpherical}
\alias{NaiveEntropySpherical,cellexalvrR-method}
\title{description of function NaiveEntropySpherical}
\usage{
NaiveEntropySpherical(x, gvect, drc, n = 20, sumFunc = sum())
}
\arguments{
\item{x}{the cellexalvrR object}

\item{gvect}{a vector with group ids}

\item{drc}{the dimension reduction name to process}

\item{n}{the distances to use (default= NULL; max dist /100 to maxdist/10 in 100 steps )}

\item{sumFunc}{function to sum up the single entropy values (default sum)}
}
\description{
description of function NaiveEntropySpherical
}
\details{
Calculates a simple entropy value for a grouping vector and a 3D drc data.
}
