% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{methods}
\name{newFilters}
\alias{newFilters}
\alias{newFilters,character-method}
\title{internal function!}
\usage{
newFilters(oldFiles, path = "../")
}
\arguments{
\item{oldFiles}{a vector of old files.}

\item{path}{the path to look for file (default '../Screenshots' )}
}
\description{
Check the folder '../' for .fil files
}
\keyword{server}
