% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{methods}
\name{newScreenshots}
\alias{newScreenshots}
\alias{newScreenshots,character-method}
\title{internal function!}
\usage{
newScreenshots(oldFiles, path = "../Screenshots/")
}
\arguments{
\item{oldFiles}{a vector of old files.}

\item{path}{the path to look for file (default '../Screenshots' )}
}
\description{
Check the folder '../Screenshots' for files
}
\keyword{server}
