% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyGOIs.R
\docType{methods}
\name{onlyGOIs}
\alias{onlyGOIs}
\alias{onlyGOIs,cellexalvrR-method}
\title{description of function onlyGOIs}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{name}{the name of the GIO list (eg TFs or epigenetic)}

\item{cellexalObj, }{TEXT MISSING}

\item{name}{TEXT MISSING}
}
\description{
Allows the user to select only (G)enes (O)f (I)nterest lists from the object
}
