% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyGOIs.R
\name{onlyGOIs}
\alias{onlyGOIs}
\title{Allows the user to select only (G)enes (O)f (I)nterest lists from the object}
\usage{
onlyGOIs(cellexalObj, name)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{name}{the name of the GIO list (eg TFs or epigenetic)}
}
\description{
Allows the user to select only (G)enes (O)f (I)nterest lists from the object
}
