% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologyLogPage.R
\docType{methods}
\name{ontologyLogPage}
\alias{ontologyLogPage}
\alias{ontologyLogPage,cellexalvrR-method}
\title{old log function to create a pathways analysis for a list of genes.}
\usage{
ontologyLogPage(
  cellexalObj,
  genes,
  grouping = NULL,
  ontology = "BP",
  topNodes = 10,
  ...
)

\S4method{ontologyLogPage}{cellexalvrR}(
  cellexalObj,
  genes,
  grouping = NULL,
  ontology = "BP",
  topNodes = 10,
  ...
)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{genes}{a list of gene symbols (IMPORTANT)}

\item{grouping}{the grouping this gene list originated on (default = NULL; use last grouping)}

\item{ontology}{which GO ontology to choose from (default = "BP")}

\item{topNodes}{how many GO terms to report (default 10)}

\item{...}{unused}
}
\description{
creates the GO analysis for a gene list and puts it into the report.
}
\details{
The ontology analysis for the log files.
Response time is too slow for VR.
}
