% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optGroupCountKmeans.R
\docType{methods}
\name{optGroupCountKmeans}
\alias{optGroupCountKmeans}
\alias{optGroupCountKmeans,cellexalvrR-method}
\title{description of function optGroupCountKmeans}
\arguments{
\item{dat}{TEXT MISSING}

\item{k}{TEXT MISSING default=1:20}
}
\description{
calculate the optimal amount of kmeans grooups for a data matrix
}
