% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optGroupCountKmeans.R
\docType{methods}
\name{optGroupCountKmeans}
\alias{optGroupCountKmeans}
\alias{optGroupCountKmeans,matrix-method}
\title{identify the optimal numer of clusters for a matrix}
\usage{
optGroupCountKmeans(dat, k = 1:20)

\S4method{optGroupCountKmeans}{matrix}(dat, k = 1:20)
}
\arguments{
\item{dat}{the data matrix}

\item{k}{number of clusters to testdefault=1:20}
}
\description{
calculate the optimal amount of kmeans grooups for a data matrix
}
\details{
Using kmeans to group the drc data
for the usability of the log file the genes need to be ordered.
The heatmap needs to show these clusters of genes. And to identify the right number of clusters I need and elbow analysis.
https://www.icsi.berkeley.edu/icsi/node/4806
Finding a Kneedle in a Haystack: Detecting Knee Points in System Behavior
Satopaa, V.., Albrecht J., Irwin D., & Raghavan B., 2010
}
