% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscilatingGenes.R
\name{oscilatingGenes}
\alias{oscilatingGenes}
\title{Identifies oscilating genes and tries to group them in oscilating clusters using the EMD package.}
\usage{
oscilatingGenes(cellexalObj, cellidfile)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}
}
\value{
Undefined at the moment
}
\description{
Identifies oscilating genes and tries to group them in oscilating clusters using the EMD package.
}
\keyword{heatmap}
