% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscilliatingGenes.R
\docType{methods}
\name{oscilliatingGenes}
\alias{oscilliatingGenes}
\alias{oscilliatingGenes,cellexalvr-method}
\title{description of function oscilliatingGenes}
\usage{
oscilliatingGenes(cellexalObj, cellidfile)
}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{cellidfile}{the grouping or a file to read a grouping from}
}
\value{
the cellexal object
}
\description{
Identifies oscilliating genes and tries to group them in oscilliating clusters using
 the EMD package. This function should not be used!
}
