% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscilliatingGenes.R
\name{oscilliatingGenes}
\alias{oscilliatingGenes}
\title{Identifies oscilliating genes and tries to group them in oscilliating clusters using the EMD package.}
\usage{
oscilliatingGenes(cellexalObj, cellidfile)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}
}
\value{
Undefined at the moment
}
\description{
Identifies oscilliating genes and tries to group them in oscilliating clusters using the EMD package.
}
\keyword{heatmap}
