% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateParts.R
\docType{methods}
\name{partFname}
\alias{partFname}
\alias{partFname,character-method}
\title{create the filename fir a cellexalvrR slot only save}
\usage{
partFname(
  part = c("meta.cell", "meta.gene", "userGroups", "usedObj", "groupSelectedFrom",
    "colors", "lastGroup", "all"),
  path
)

\S4method{partFname}{character}(
  part = c("meta.cell", "meta.gene", "userGroups", "usedObj", "groupSelectedFrom",
    "colors", "lastGroup", "all"),
  path
)
}
\arguments{
\item{part}{to get the filename for (  'meta.cell',  'meta.gene',  'userGroups',  'usedObj', 'all' )}

\item{path}{the outpath}
}
\description{
returns the possible outfile for the partial saving of the cellexal object
}
\details{
The function partFname checks which part to save and throws an error if the part is not avaialable for saving.
}
