% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\docType{methods}
\name{plotAllMDS}
\alias{plotAllMDS}
\alias{plotAllMDS,cellexalvrR-method}
\title{description of function plotAllMDS}
\arguments{
\item{cellexalObj, }{cellexalvr object}
}
\description{
Plots the MDS reduced data for a quick look
}

