% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotAllMDS}
\alias{plotAllMDS}
\title{Plots the MDS reduced data for a quick look}
\usage{
plotAllMDS(cellexalObj)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}
}
\description{
Plots the MDS reduced data for a quick look
}
