% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{plotDataOnTime,cellexalTime,list-method}
\alias{plotDataOnTime,cellexalTime,list-method}
\title{Plot a list of data values on the timeline the vectors in the list must have cell names}
\usage{
\S4method{plotDataOnTime}{cellexalTime,list}(x, dat, color = NULL, ofile, smooth = TRUE)
}
\description{
Plot a list of data values on the timeline the vectors in the list must have cell names
}
