% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{plotDataOnTime,data.frame,list-method}
\alias{plotDataOnTime,data.frame,list-method}
\title{Plot a list of data values on a data.frame containing the time information
The vectors in the list must have cell names}
\usage{
\S4method{plotDataOnTime}{data.frame,list}(x, dat, color = NULL, ofile, smooth = TRUE)
}
\description{
Plot a list of data values on a data.frame containing the time information
The vectors in the list must have cell names
}
