% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{plotDataOnTime}
\alias{plotDataOnTime}
\alias{plotDataOnTime,cellexalTime,list-method}
\alias{plotDataOnTime,data.frame,list-method}
\title{plot data on the cellexalTime time information as XY plots}
\usage{
plotDataOnTime(x, dat, color = NULL, ofile, smooth = TRUE)

\S4method{plotDataOnTime}{cellexalTime,list}(x, dat, color = NULL, ofile, smooth = TRUE)

\S4method{plotDataOnTime}{data.frame,list}(x, dat, color = NULL, ofile, smooth = TRUE)
}
\arguments{
\item{x}{the cellexalTime object or a data.frame (use the cellexalTime object)}

\item{dat}{the linear data to plot}

\item{color}{the color(s) the data should be plotted in}

\item{ofile}{the figure file (png)}

\item{smooth}{add a smoothed line to the graph (default TRUE)}
}
\description{
plot linear data on the time in a 2d figure
}
\details{
Plot a list of data values on the timeline the vectors in the list must have cell names
}
