% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{plotDataOnTime_rmd,cellexalTime-method}
\alias{plotDataOnTime_rmd,cellexalTime-method}
\title{Here the plot png files are converted to Rmd links}
\usage{
\S4method{plotDataOnTime_rmd}{cellexalTime}(x, dat, color = NULL, ofile, cellexalObj)
}
\description{
Here the plot png files are converted to Rmd links
}
