% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEntropyDiff_2D_3D.R
\docType{methods}
\name{plotEntropyDiff_2D_3D}
\alias{plotEntropyDiff_2D_3D}
\alias{plotEntropyDiff_2D_3D,cellexalvrR-method}
\title{description of function plotEntropyDiff_2D_3D}
\usage{
plotEntropyDiff_2D_3D(entro2D, entro3D, file, size = 800)
}
\arguments{
\item{entro2D}{a 2D entropy calculation}

\item{entro3D}{2 3D entropy calculation}

\item{file}{the plot outfile base name}

\item{size}{the hight and width of the png files default=800}
}
\description{
plots the dcr entropy as shown in the paper
}
\details{
plots the first row of a table into two png xy line plots based 
on two tables in double log scale (1) and scaled to the last value in row 2 (2)
Iedally these tables have been produced by a call to NaiveEntropySpherical to qualify the entropy of a 2D and 3D mds plot.
}
