% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03cellexalGrouping.R
\docType{methods}
\name{plotGroup}
\alias{plotGroup}
\alias{plotGroup,cellexalvrR,cellexalGrouping-method}
\title{show for a cellexalGrouping object}
\usage{
plotGroup(y, x, gname = NULL, showIDs = TRUE)

\S4method{plotGroup}{cellexalvrR,cellexalGrouping}(y, x, gname = NULL, showIDs = TRUE)
}
\arguments{
\item{y}{the cellexalvrR object}

\item{x}{the cellexalGrouping}

\item{gname}{the gene name to plot expression for (default NULL)}

\item{showIDs}{show the group IDs in the plot ( default = TRUE)}
}
\description{
plotGroup the cellexalGrouping
}
