% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,cellexalvrR-method}
\title{description of function plotHeatmap}
\usage{
plotHeatmap(
  x,
  genes,
  groupings,
  ofile = NULL,
  width = 9,
  height = 9,
  main = NULL,
  X11type = "cairo",
  family = "Helvetica",
  fileType = "pdf"
)
}
\arguments{
\item{genes}{the genes to include into the heatmap (correct order!)}

\item{groupings}{either columns in the userGroups table
or a list of CellexalVR selection files or a mix of both}

\item{ofile}{the figure out file (or null)}

\item{width}{default figure width (9 in)}

\item{height}{default figure height (9 in)}

\item{main}{the figure titla (defaults to gene name)}

\item{X11type}{not important here (default to 'cairo')}

\item{family}{pdf typing family (defaults to "Helvetica") png is also supported}

\item{fileType}{figure file type (default 'pdf' )}
}
\description{
Create a simple violin plot on the expresion of one gene
}
\details{
Plot a simple violin plot using the cellexalVR colors
}
