% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{plotHeatmap_rmd,matrix-method}
\alias{plotHeatmap_rmd,matrix-method}
\title{convert the pngs from plotTimeHeatmap to Rmd figures.}
\usage{
\S4method{plotHeatmap_rmd}{matrix}(x, ofile, cellexalObj, color = NULL)
}
\description{
convert the pngs from plotTimeHeatmap to Rmd figures.
}
