% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{plotHeatmap_rmd}
\alias{plotHeatmap_rmd}
\alias{plotHeatmap_rmd,cellexalTime-method}
\title{description of function plot}
\arguments{
\item{x}{the matrix to plot (genes rows cells cols)}

\item{ofile}{the figure file (png)}

\item{cellexalObj}{the cellexal object to adjust the path for the markdown file.}

\item{color}{the color(s) the data should be plotted in}
}
\description{
plot linear data on the time in a 2d figure
}
