% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{plotHeatmap_rmd}
\alias{plotHeatmap_rmd}
\alias{plotHeatmap_rmd,matrix-method}
\title{clip the figure path for using in the log system}
\usage{
plotHeatmap_rmd(x, ofile, cellexalObj, color = NULL)

\S4method{plotHeatmap_rmd}{matrix}(x, ofile, cellexalObj, color = NULL)
}
\arguments{
\item{x}{the matrix to plot (genes rows cells cols)}

\item{ofile}{the figure file (png)}

\item{cellexalObj}{the cellexal object to adjust the path for the markdown file.}

\item{color}{the color(s) the data should be plotted in}
}
\description{
plot linear data on the time in a 2d figure
}
\details{
convert the pngs from plotTimeHeatmap to Rmd figures.
}
