% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{plotTime}
\alias{plotTime}
\alias{plotTime,cellexalTime-method}
\title{plot the drc with time colors}
\usage{
plotTime(x)

\S4method{plotTime}{cellexalTime}(x)
}
\arguments{
\item{x}{the object}
}
\description{
3D plot the time with a line in the time
}
\details{
Plot the linear data with the correct colors.
Used to debug the linear selections.
}
