% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\name{plotTimeHeatmap,matrix-method}
\alias{plotTimeHeatmap,matrix-method}
\title{<circle cx="50" cy="50" r="50"/>',}
\usage{
\S4method{plotTimeHeatmap}{matrix}(x, ofile, color = NULL, circleF = NULL)
}
\description{
<circle cx="50" cy="50" r="50"/>',
}
