% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{plotTimeHeatmap}
\alias{plotTimeHeatmap}
\alias{plotTimeHeatmap,cellexalTime-method}
\title{description of function plot}
\arguments{
\item{x}{the zscored data object}

\item{ofile}{the figure file (a '.png' will be added)}

\item{color}{the color of the border around the heatmap}
}
\value{
the figure file path
}
\description{
create a really simple heatmap of the zscored data
}
