% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{plotTimeHeatmap}
\alias{plotTimeHeatmap}
\alias{plotTimeHeatmap,matrix-method}
\title{description of function plotTimeHeatmap}
\usage{
plotTimeHeatmap(x, ofile, color = NULL, circleF = NULL)

\S4method{plotTimeHeatmap}{matrix}(x, ofile, color = NULL, circleF = NULL)
}
\arguments{
\item{x}{the zscored data object}

\item{ofile}{the figure file (a '.png' will be added)}

\item{color}{the color of the border around the heatmap}

\item{circleF}{a file for a small colored circle as svg file (needed for the reports)}
}
\value{
create a heatmap figure file for a cellexalTime object
}
\description{
create a really simple heatmap of the zscored data
}
