% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\docType{methods}
\name{plotViolin}
\alias{plotViolin}
\alias{plotViolin,cellexalvrR-method}
\title{create violin plots for a list of genes and a grouping (interactive only)}
\usage{
plotViolin(
  x,
  gene,
  grouping,
  ofile = NULL,
  width = 9,
  height = 9,
  main = NULL,
  X11type = "cairo",
  family = "Helvetica"
)

\S4method{plotViolin}{cellexalvrR}(
  x,
  gene,
  grouping,
  ofile = NULL,
  width = 9,
  height = 9,
  main = NULL,
  X11type = "cairo",
  family = "Helvetica"
)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{gene}{the gene name}

\item{grouping}{either a column in the userGroups table or a CellexalVR selection file}

\item{ofile}{the figure out file (or null)}

\item{width}{default figure width (9 in)}

\item{height}{default figure height (9 in)}

\item{main}{the figure title (defaults to gene name)}

\item{X11type}{not important here (default to 'cairo')}

\item{family}{pdf typing family (defaults to "Helvetica") png is also supported}
}
\description{
Create a simple violin plot on the expresion of one gene
}
\details{
Plot a simple violin plot using the cellexalVR colors
}
