% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPlot2Dtime.R
\docType{methods}
\name{prettyPlot2Dtime}
\alias{prettyPlot2Dtime}
\alias{prettyPlot2Dtime,cellexalvrR-method}
\title{description of function prettyPlot2Dtime}
\usage{
prettyPlot2Dtime(x)
}
\arguments{
\item{x}{the data.frame containing the data for the plot}
}
\description{
A rather naive ggplot2 function to plot a simple Matrix.
}
