% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_Seurat_ATAC.R
\docType{methods}
\name{process_Seurat_ATAC_combined}
\alias{process_Seurat_ATAC_combined}
\alias{process_Seurat_ATAC_combined,environment-method}
\title{import a Seurat atac combined analysis}
\usage{
process_Seurat_ATAC_combined(
  rna,
  atac,
  merged,
  meta.cell.groups = NULL,
  outpath = getwd(),
  specie
)
}
\arguments{
\item{rna}{the Seurat rna analysis object}

\item{atac}{the Seurat atac analysis object}

\item{merged}{the seurat combined analysis object}

\item{meta.cell.groups}{which meta.data columns to convert to meta.cell classes}

\item{outpath}{set the outpath of the object (default getwd())}

\item{specie}{set the specie to either mouse or human (default check gene names)}
}
\description{
convert a BioData list (BioData library not loaded) into a cellexalvrR obejct
}
\details{
The aim of this function is to make as much data available for CellexalVR as possible.
3D models for all 3 analyses and the corresponding data will be added.
}
