% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVelocytoPyResults.R
\docType{methods}
\name{processVelocytoPyResults}
\alias{processVelocytoPyResults}
\alias{processVelocytoPyResults,cellexalvrR-method}
\title{description of function processVelocytoPyResults}
\usage{
processVelocytoPyResults(
  total,
  variable,
  embeddings = c("umap", "phate"),
  embeddingDims = 3,
  velocyto = TRUE,
  veloScale = 20,
  specie = "human",
  minCell4gene = 10
)
}
\arguments{
\item{total}{the hdf5 file containing all expression values (normalized)}

\item{variable}{the hdf5 file containing the embeddings and velocyto data}

\item{embeddings}{the embedding names default= c('umap', 'phate')}

\item{embeddingDims}{the dimensionality of the embeddings (default 3)}

\item{velocyto}{should velocyto data be added to the embeddings default=TRUE}

\item{veloScale}{velocyto scaling factor default=20}

\item{specie}{the species the data has been created from default='human'}

\item{minCell4gene}{the total file might contain genes not expressed at all - filter them default= 10}
}
\description{
Combine two anndata files into one cellexalvrR object
}
\details{
Velocytopy is becoming more and more important and the analysis (normally)
creates a normalized expression hdf5 file and a more detailed (e.g. variable genes based)
dimensional reduction hdf5 file.

The processed file likely also has velocity information which can also be added 
in the correct way to the cellexalvrR object.

For the VR process bothe the embedings as well as (all) expression data
is of interest.

Here both files are processed and the resulting
}
