% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTo.R
\docType{methods}
\name{reduceTo}
\alias{reduceTo}
\alias{reduceTo,cellexalvr-method}
\title{description of function reduceTo}
\usage{
reduceTo(cellexalvr, what = "row", to = NULL, ...)
}
\arguments{
\item{what}{reduce to meta.cel or row ids default='row'
to select these names default=NULL}

\item{x}{the NGScollation object}

\item{name}{the new name of this object (mainly unused)}
}
\description{
The main reduction function can drop both meta.cel and genes using the colnames / rownames of the data tables
}
