% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderFile.R
\docType{methods}
\name{renderFile}
\alias{renderFile}
\alias{renderFile,cellexalvrR-method}
\title{description of function renderFile}
\usage{
renderFile(x, id)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{id}{the id of the report file to render ('x@usedObj$sessionRmdFiles[id]')}
}
\description{
render only one html section, not the whole session log
}
