% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderReport.R
\docType{methods}
\name{renderReport,character-method}
\alias{renderReport,character-method}
\title{description of function renderReport}
\usage{
\S4method{renderReport}{character}(cellexalObj)
}
\arguments{
\item{cellexalObj}{the cellexal.RData file}
}
\description{
preload the cellexalObj
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: cellexalvrR
}}

