% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renew.R
\docType{methods}
\name{renew}
\alias{renew}
\alias{renew,cellexalvrR-method}
\title{description of function renew}
\usage{
\S4method{renew}{cellexalvrR}(x)
}
\arguments{
\item{x}{the object you want to updatae}
}
\description{
updatae the class definition by re-creating the instance This version also makes sure,

that the returned object is an S4 object.
}
