% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renew.R
\docType{methods}
\name{renew,cellexalvr-method}
\alias{renew,cellexalvr-method}
\title{description of function renew}
\usage{
\S4method{renew}{cellexalvr}(x)
}
\arguments{
\item{x}{the old cellexalvr (not cellexalvrR) object}
}
\description{
renew the OLD cellexalvr objects
}
\section{Functions}{
\itemize{
\item \code{renew,cellexalvr-method}: cellexalvrR
}}

