% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.genes.R
\docType{methods}
\name{reorder.genes}
\alias{reorder.genes}
\alias{reorder.genes,cellexalvr-method}
\title{description of function remove.genes}
\arguments{
\item{dataObj}{the cellexalvr object}

\item{column}{the samples column to reorder on}
}
\description{
this function reorderes the cellexalvr object based on a column in the samples table

(e.g. for plotting)
}
