% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.obj.R
\docType{methods}
\name{reorder.samples}
\alias{reorder.samples}
\alias{reorder.samples,cellexalvr-method}
\title{description of function remove.genes}
\usage{
\method{reorder}{samples}(dataObj, column)
}
\arguments{
\item{dataObj}{the cellexalvr object}

\item{column}{the annotation column to reorder on}
}
\description{
this function reorderes the cellexalvr object based on a column in the annotation table (e.g. for plotting)
}
