% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.obj.R
\docType{methods}
\name{reorderGenes}
\alias{reorderGenes}
\alias{reorderGenes,cellexalvrR-method}
\title{simple reordering of the genes}
\usage{
reorderGenes(dataObj, column)

\S4method{reorderGenes}{cellexalvrR}(dataObj, column)
}
\arguments{
\item{dataObj}{the cellexalvr object}

\item{column}{the samples column to reorder on}
}
\description{
this function reorderes the cellexalvr object based on a column in the samples table

(e.g. for plotting)
}
\details{
reorder function to reorder the whole cellexalvrR object based on a genes order.
}
