% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.obj.R
\docType{methods}
\name{reorderSamples}
\alias{reorderSamples}
\alias{reorderSamples,cellexalvrR-method}
\title{simple reordering of the samples}
\usage{
reorderSamples(dataObj, column)

\S4method{reorderSamples}{cellexalvrR}(dataObj, column)
}
\arguments{
\item{dataObj}{the cellexalvr object}

\item{column}{the annotation column to reorder on}
}
\description{
this function reorderes the cellexalvr object based on a column in the annotation

table (e.g. for plotting)
}
\details{
reorder function to reorder the whole cellexalvrR object based on a cell order.
}
