% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\docType{methods}
\name{reset}
\alias{reset}
\alias{reset,cellexalvrR-method}
\title{description of function reset}
\usage{
\S4method{reset}{cellexalvrR}(x)
}
\arguments{
\item{x}{the cellexalObj you want to reset}
}
\description{
remove all internals originating from previouse analyses
}
