% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\docType{methods}
\name{reset}
\alias{reset}
\alias{reset,cellexalvrR-method}
\title{remove all modifications added to this cellexalvrR object}
\usage{
reset(x)

\S4method{reset}{cellexalvrR}(x)
}
\arguments{
\item{x}{the cellexalObj you want to reset}
}
\description{
remove all internals originating from previouse analyses
}
\details{
Remove all modifications produced through VR run.
}
