% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdLink.R
\docType{methods}
\name{rmdLink}
\alias{rmdLink}
\alias{rmdLink,character-method}
\title{easily create an Rmd link that opens in a new window.}
\usage{
rmdLink(name, link, lineEnd = T)

\S4method{rmdLink}{character}(name, link, lineEnd = T)
}
\arguments{
\item{name}{the displayed text}

\item{link}{the link address}

\item{lineEnd}{add a line end at the end of every entry (default =T)}
}
\description{
creates a linke in the structure [<name>](<link><name>){target='blank'}
}
\details{
Create a R markdown link used in the (VR) log files.
}
