% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptHelpers.R
\name{Rscript.exe}
\alias{Rscript.exe}
\title{create the ''' wrapped location of the Rscript/Rscript.exe program}
\usage{
Rscript.exe()
}
\description{
create the ''' wrapped location of the Rscript/Rscript.exe program
}
