% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.ddrtree.R
\docType{methods}
\name{run.ddrtree}
\alias{run.ddrtree}
\alias{run.ddrtree,cellexalvr-method}
\title{description of function run.ddrtree}
\arguments{
\item{seuratObj}{a seurat object}
}
\description{
creates a DDRTree 3D matrix from a seurat object using only the variable genes.
}
