% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateParts.R
\docType{methods}
\name{savePart}
\alias{savePart}
\alias{savePart,cellexalvrR-method}
\title{description of function savePart}
\arguments{
\item{x}{TEXT MISSING}

\item{part}{TEXT MISSING default= c( 'meta.cell'}

\item{path}{the optional outpath (default x@outpath)}
}
\description{
save only a part of the cellexal object increasing the VR interaction speed
}
