% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateParts.R
\docType{methods}
\name{savePart}
\alias{savePart}
\alias{savePart,cellexalvrR-method}
\title{save only a slot of the cellexalObj}
\usage{
savePart(
  x,
  part = c("meta.cell", "meta.gene", "userGroups", "usedObj", "groupSelectedFrom",
    "colors", "lastGroup"),
  path = NULL
)

\S4method{savePart}{cellexalvrR}(
  x,
  part = c("meta.cell", "meta.gene", "userGroups", "usedObj", "groupSelectedFrom",
    "colors", "lastGroup"),
  path = NULL
)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{part}{one of tge slots in cellexalvrR}

\item{path}{an optional outpath (default x@outpath)}
}
\description{
save only a part of the cellexal object increasing the VR interaction speed
}
\details{
The function savePart is one main part in the load and save speed up implemented for cellexalvrR

Instead of saving all data in each VR script call only the modified data is saved.
This function saves a specific data part of the whole cellexalvrR speeding up the save process.
}
