% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{methods}
\name{server}
\alias{server}
\alias{server,character-method}
\title{start a server function periodicly sourcing in a file.}
\usage{
server(file, sleepT = 1, debug = FALSE, masterPID = NULL, asFunction = FALSE)

\S4method{server}{character}(file, sleepT = 1, debug = FALSE, masterPID = NULL, asFunction = FALSE)
}
\arguments{
\item{file}{the file core name to create input.R input.log and pid files.}

\item{sleepT}{sleep time in seconds between checks for the input.R file (default 1)}

\item{debug}{create the server output file? default FALSE}

\item{masterPID}{if this pid is not active any more stop the server (default NULL)}

\item{asFunction}{do not shut down R when exiting (default FALSE)}
}
\description{
starts a server looking for a sourcable file in paste(file, 'input.R', sep='.')
The function sets a paste(file, 'input.lock', sep='.') lock file on processed scripts.
To shut down the server process you can either write a q('no') into the script file or remove the pid file.
}
\details{
This server function is the main speed up for big data files.

Instead of loading and saving all files for each function call in the VR process
we now can have one R process that sequencially works on all requests.

The server will look for a file called <file>.input.R and process all commands in the file 
using source(fname) so please be careful there!

While the file <file>.input.lock is existing the input.R file is processed by the R and should not be touched.
}
\keyword{server}
