% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionCounter.R
\name{sessionCounter}
\alias{sessionCounter}
\alias{sessionCounter,cellexalvrR-method}
\title{an internal function to get the ID for the next log file}
\usage{
sessionCounter(x, gName)

\S4method{sessionCounter}{cellexalvrR}(x, gName)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{gName}{the group name}
}
\description{
returns the counter for this session and this group name
}
\details{
A log helper function. It does remove the need to store all log Rmd files in the object.
}
