% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionPath.R
\docType{methods}
\name{sessionPath}
\alias{sessionPath}
\alias{sessionPath,cellexalvrR-method}
\title{description of function sessionPath}
\usage{
sessionPath(cellexalObj, sessionName = NULL)

\S4method{sessionPath}{cellexalvrR}(cellexalObj, sessionName = NULL)

\S4method{sessionPath}{character}(cellexalObj, sessionName = NULL)
}
\arguments{
\item{cellexalObj}{the cellexal object}

\item{sessionName}{the session ID default=NULL}
}
\description{
Use the session ID and object outpath to create a session path for the reports
}
\details{
VR method that creates the session specific settings for each VR session.
}
