% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionPath.R
\docType{methods}
\name{sessionPath,character-method}
\alias{sessionPath,character-method}
\title{description of function sessionPath}
\usage{
\S4method{sessionPath}{character}(cellexalObj, sessionName = NULL)
}
\arguments{
\item{cellexalObj}{the cellexalOvh.RData file}

\item{sessionName}{the session ID default=NULL}
}
\description{
preload the cellexalOvh.RData file
}
\section{Functions}{
\itemize{
\item \code{sessionPath,character-method}: cellexalvrR
}}

