% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\docType{methods}
\name{set.specie}
\alias{set.specie}
\alias{set.specie,cellexalvrR-method}
\title{description of function set.specie}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{specie}{The specie required options: "mouse" or "human"}
}
\description{
Loads TF annotation into cellexalvr object
}
\details{
set.specie is depricated

It simply sets the specie slot to the user supplied specie string.
}
\keyword{TFs}
