% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\name{set.specie}
\alias{set.specie}
\title{Loads TF annotation into cellexalvr object}
\usage{
set.specie(cellexalObj, specie = c("mouse", "human"))
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{specie}{The specie required}
}
\description{
Loads TF annotation into cellexalvr object
}
\keyword{TFs}
