% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentials.R
\docType{methods}
\name{setStatsMethod}
\alias{setStatsMethod}
\alias{setStatsMethod,cellexalvrR-method}
\title{set the default stats method}
\usage{
setStatsMethod(x, method = "anova")
}
\arguments{
\item{x}{the cellexalvrR object}

\item{method}{the default stats method to use (default 'anova')}
}
\description{
sets the used statistics method from this list ( "anova","edgeR", "MAST", 'Seurat' )
}
