% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat2cellexalvr.R
\docType{methods}
\name{seurat2cellexalvr}
\alias{seurat2cellexalvr}
\alias{seurat2cellexalvr,cellexalvr-method}
\title{description of function seurat2cellexalvr}
\arguments{
\item{seuratObj}{the suerat object to be converted}

\item{seuratObj}{TEXT MISSING}
}
\description{
Converts a seurat class to one of cellexalvr
}
