% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrExportFunctions.R
\name{seurat2cellexalvr}
\alias{seurat2cellexalvr}
\title{Converts a seurat class to one of cellexalvr. If the cell-cycle score have been calculated they will be added to the cell.mata table}
\usage{
seurat2cellexalvr(seuratObj)
}
\arguments{
\item{seuratObj}{the suerat object to be converted}
}
\description{
Converts a seurat class to one of cellexalvr. If the cell-cycle score have been calculated they will be added to the cell.mata table
}
