% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03cellexalGrouping.R, R/show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,cellexalGrouping-method}
\alias{show,cellexalvrR-method}
\title{show for a cellexalGrouping object}
\usage{
\S4method{show}{cellexalGrouping}(object)
}
\arguments{
\item{object}{the cellexalvr object}
}
\value{
nothing at all
}
\description{
show the objects internals

show the cellexalvrR
}
