% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R, R/03cellexalGrouping.R,
%   R/show.R
\docType{methods}
\name{show}
\alias{show}
\title{show the object contents}
\usage{
\S4method{show}{cellexalTime}(object)

\S4method{show}{cellexalGrouping}(object)

\S4method{show}{cellexalvrR}(object)
}
\arguments{
\item{object}{the cellexalvr object}
}
\value{
nothing at all

nothing at all
}
\description{
show the cellexalvrR

show the objects internals

show the cellexalvrR
}
