% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlotHeatmaps.R
\name{simplePlotHeatmaps,cellexalvrR,list,character-method}
\alias{simplePlotHeatmaps,cellexalvrR,list,character-method}
\title{<circle cx="50" cy="50" r="50"/>',}
\usage{
\S4method{simplePlotHeatmaps}{cellexalvrR,list,character}(x, info, fname)
}
\description{
<circle cx="50" cy="50" r="50"/>',
}
