% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlotHeatmaps.R
\docType{methods}
\name{simplePlotHeatmaps}
\alias{simplePlotHeatmaps}
\alias{simplePlotHeatmaps,cellexalvrR,cellexalGrouping,character-method}
\title{create the heatmaps for the (interactive only) timline logs}
\usage{
simplePlotHeatmaps(x, info, fname)

\S4method{simplePlotHeatmaps}{cellexalvrR,cellexalGrouping,character}(x, info, fname)
}
\arguments{
\item{x, }{cellexalvrR object that has the correct order and only the genes of interest!}

\item{info}{the grouping info for this plot}

\item{fname}{the outfile base (.png for main .<i>.png for the slices)}
}
\value{
a list with the keys 
  png - all outfiles created; first summary others heatmaps),
  genes - the genes split into the displayed groups,
  ofile - the main outfile,
  error - any error occuring - should be included in the report,
  smoothedClusters - the smoothened lines,
 MaxInCluster - which line has in which time slot it's maximum,
  mat - the z-scored summed up expression matrix.
}
\description{
plot an extremely simple heatmap and slices of that
}
\details{
The log logics sometimes would benefit froma usable descriptions of the expression patterns.
Here we use the most simple heatmap functionallity in R (image) to show the expression values.
The returned genes list has the genes in the order of display in the heatmaps.

In order to add usability the hetmap is split into an optimal number of sub heatmaps to show the trends.

A new permanent variable in <cellexalvrR object>@userObj$gene_clusters (integer) can be used to override 
WSS based optimal grouping search.
}
