% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_logging.R
\docType{methods}
\name{start_logging}
\alias{start_logging}
\alias{start_logging,cellexalvr-method}
\title{description of function start_logging}
\usage{
start_logging(x)
}
\arguments{
\item{x}{TEXT MISSING}
}
\description{
creates a taskCallbackManager in x$usedObj$tcbm which hopefully loggs all additional changes
}
