% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderFile.R
\docType{methods}
\name{storeLogContents}
\alias{storeLogContents}
\alias{storeLogContents,cellexalvrR-method}
\title{description of function storeLogContents}
\usage{
storeLogContents(x, content)
}
\arguments{
\item{x}{the CellexalvrR object}

\item{content}{the text content of the Rmd file}
}
\description{
store the Rmd contens produced in a logXYZ function into a file and register that one.
}
