% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderFile.R
\docType{methods}
\name{storeLogContents}
\alias{storeLogContents}
\alias{storeLogContents,cellexalvrR-method}
\title{write a (Rmd) string and add that to the log}
\usage{
storeLogContents(x, content, type = "")

\S4method{storeLogContents}{cellexalvrR}(x, content, type = "")
}
\arguments{
\item{x}{the CellexalvrR object}

\item{content}{the text content of the Rmd file}

\item{type}{the type of log saved (default '')}
}
\description{
store the Rmd contens produced in a logXYZ function into a file and register that one.
}
