% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{subset}
\alias{subset}
\title{subset a cellexalLinear object}
\usage{
\S4method{subset}{cellexalLinear}(x, cells)
}
\arguments{
\item{x}{the object}

\item{cells}{the cell names to subset the object to}
}
\description{
subset a cellexalLinear object
}
\details{
Subset a cellexalLinear object based on a list of cells.
}
