% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{subsetTime}
\alias{subsetTime}
\alias{subsetTime,cellexalTime-method}
\title{description of function plot}
\usage{
subsetTime(x, cells)
}
\arguments{
\item{x}{the object}

\item{cells}{the cell names to subset the time to}
}
\description{
3D plot the time with a line in the time
}
