% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{subsetTime}
\alias{subsetTime}
\alias{subsetTime,cellexalTime-method}
\title{subset a cellexalTime object}
\usage{
subsetTime(x, cells)

\S4method{subsetTime}{cellexalTime}(x, cells)
}
\arguments{
\item{x}{the object}

\item{cells}{the cell names to subset the time to}
}
\description{
subset a cellexalTime object
}
\details{
Subset a time object based on a list of cells.
}
