% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{timeAnalysisSubset}
\alias{timeAnalysisSubset}
\alias{timeAnalysisSubset,cellexalTime-method}
\title{old method combining createStats and createReport functions}
\usage{
timeAnalysisSubset(x, cellexalObj, deg.genes = NULL)

\S4method{timeAnalysisSubset}{cellexalTime}(x, cellexalObj, deg.genes = NULL)
}
\arguments{
\item{x}{the object}

\item{cellexalObj}{the cellexal object to process}

\item{deg.genes}{dummary over these genes of the cellexalObj@usedObj$deg.genes populated during the createStats() run}
}
\description{
create a time report for a subset of cells only
}
\details{
To run this method efficiently it is recommended to first subset the drc the selection is based on.
Next it is recommended to add this drc subset to the cellexal object and use the subsetTime function to also subset the timeline.
This way one can assure that the timepoints can be compared across the analyses.
This function will run the statistics and create a report for the new subset.

This function is highly experimental and not meant to be used from within VR!
}
