% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{toString}
\alias{toString}
\title{stringify the timeline (print/show)}
\usage{
\S4method{toString}{cellexalTime}(x)
}
\arguments{
\item{x}{the cellexalTime object}
}
\description{
summarize a cellexalTime object as string
}
\details{
Intern method to convert the object to a string for the show function.
}
