% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useInbuiltGOIlists.R
\docType{methods}
\name{useInbuiltGOIlists}
\alias{useInbuiltGOIlists}
\alias{useInbuiltGOIlists,cellexalvr-method}
\title{description of function useInbuiltGOIlists}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{name}{the name of the inbuilt list to use ( either 'TFs' or 'epigenetic' for now)}

\item{cellexalObj}{TEXT MISSING}

\item{name}{TEXT MISSING}
}
\description{
An easy function to register the inbuilt (G)enes (O)f (I)nterest lists 'TFs' and

'epigenetic' are supported at the moment
}
