% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useInbuiltGOIlists.R
\name{useInbuiltGOIlists}
\alias{useInbuiltGOIlists}
\title{An easy function to register the inbuilt (G)enes (O)f (I)nterest lists 'TFs' and 'epigenetic'}
\usage{
useInbuiltGOIlists(cellexalObj, name)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{name}{the name of the inbuilt list to use ( either 'TFs' or 'epigenetic' for now)}
}
\description{
An easy function to register the inbuilt (G)enes (O)f (I)nterest lists 'TFs' and 'epigenetic'
}
