% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userGrouping.R
\docType{methods}
\name{userGrouping}
\alias{userGrouping}
\alias{userGrouping,cellexalvr-method}
\title{description of function userGrouping}
\usage{
userGrouping(cellexalObj, cellidfile)
}
\arguments{
\item{cellexalObj}{TEXT MISSING}

\item{cellidfile}{TEXT MISSING}
}
\description{
Reads a VR grouping file and creates a user.grouping column with the information

storing the user defined grouing for later use
}
