% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userGrouping.R
\name{userGrouping}
\alias{userGrouping}
\title{Reads a VR grouping file and creates a user.grouping column with the information
storing the user defined grouing for later use}
\usage{
userGrouping(cellexalObj, cellidfile)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}
}
\description{
Reads a VR grouping file and creates a user.grouping column with the information
storing the user defined grouing for later use
}
\keyword{userGrouping}
