% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_vioplot.R
\docType{methods}
\name{vioplot}
\alias{vioplot}
\alias{vioplot,Rscexv-method}
\title{description of function vioplot}
\usage{
vioplot(
  datas,
  range = 1.5,
  h = NULL,
  ylim = NULL,
  names = NULL,
  horizontal = FALSE,
  col = "magenta",
  border = "black",
  lty = 1,
  lwd = 1,
  rectCol = "black",
  colMed = "white",
  pchMed = 19,
  at,
  add = FALSE,
  wex = 1,
  drawRect = TRUE,
  main = NULL,
  cex.axis = 1,
  neg = NULL
)
}
\arguments{
\item{datas}{a list with all data to plot}

\item{range}{see orig vioplot funciton default= 1.5}

\item{h}{see orig vioplot funciton default= NULL}

\item{ylim}{see orig vioplot funciton default= NULL}

\item{names}{see orig vioplot funciton default= NULL}
}
\description{
the vioplot function patched to allow different colors for the different violines.
}
\details{
This function is a hack of the official vioplot function to allow multiple colors of the violines.
}
