% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRmode.R
\name{VRmode,ANY-method}
\alias{VRmode,ANY-method}
\title{A function checking whether VR is run at the same time as the R process.}
\usage{
\S4method{VRmode}{ANY}(file, sleepT = 1, debug = FALSE, masterPID = NULL, asFunction = FALSE)
}
\description{
A function checking whether VR is run at the same time as the R process.
}
