% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRmode.R
\docType{methods}
\name{VRmode}
\alias{VRmode}
\alias{VRmode,character-method}
\title{check if CellexalVR.exe is running (Windows specific)}
\usage{
VRmode(file, sleepT = 1, debug = FALSE, masterPID = NULL, asFunction = FALSE)
}
\value{
boolean value if the VR application is running.
}
\description{
Windows specific check if CellexalVR.exe is running
}
\keyword{VRmode}
