% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRmode.R
\docType{methods}
\name{VRmode}
\alias{VRmode}
\title{check if CellexalVR.exe is running (Windows specific)}
\usage{
VRmode()
}
\value{
boolean value if the VR application is running.
}
\description{
Windows specific check if CellexalVR.exe is running
}
\keyword{VRmode}
