% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrPrepare.R
\docType{methods}
\name{vrPrepare}
\alias{vrPrepare}
\alias{vrPrepare,cellexalvr-method}
\title{description of function vrPrepare}
\usage{
vrPrepare(x, max = 10)
}
\arguments{
\item{x}{TEXT MISSING}

\item{max}{the maximum complexity of one column to split it into max 0/1 columns (default =10)}

\item{max}{TEXT MISSING default=10}
}
\value{
the updated cellexalvr object
}
\description{
Converts all sample annoation into 0/1 columns with meaningful names and adds all

inbuilt gene annoations to the meta.gene table for later use.
}
