% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportFunctions.R
\docType{methods}
\name{write_as_sqlite3}
\alias{write_as_sqlite3}
\alias{write_as_sqlite3,cellexalvrR-method}
\title{write a cellexalvrR objects data to a 'sqlite3' database of name ofile.}
\usage{
write_as_sqlite3(cellexalObj, ofile)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{ofile}{the database outfile}
}
\description{
save the cellexalObj@data object without questions asked.
}
