% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportFunctions.R
\docType{methods}
\name{write_as_sqlite3}
\alias{write_as_sqlite3}
\alias{write_as_sqlite3,cellexalvrR-method}
\title{write a cellexalvrR object's data to a 'sqlite3' database of name ofile.}
\usage{
write_as_sqlite3(x, ofile)

\S4method{write_as_sqlite3}{cellexalvrR}(x, ofile)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{ofile}{the database outfile}
}
\description{
save the x@data object without questions asked.
}
\details{
SQLite is the database the VR reads the expression values from.

This function creates the databases from the data stored in the data slot.
}
