% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\docType{methods}
\name{addCellMeta2cellexalvr}
\alias{addCellMeta2cellexalvr}
\alias{addCellMeta2cellexalvr,cellexalvrR-method}
\title{description of function 'addCellMeta2cellexalvr'}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{cell.meta}{A matrix of cell metadata}
}
\description{
Adds per cell metadata to a 'cellexalvrObj'
}
\details{
This function adds a 0/1 information matrix to the cellexalvrR object.

If the sample annotation data is not in 0/1 format please convert it using
make.cell.meta.from.df(). #function definition in file 'make.cell.meta.from.df.R'
}
