% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addElements.R
\docType{methods}
\name{addFACS2cellexalvr}
\alias{addFACS2cellexalvr}
\alias{addFACS2cellexalvr,cellexalvrR-method}
\title{Adds FACS index to a cellexalvrObj}
\usage{
addFACS2cellexalvr(cellexalObj, facs.data)

\S4method{addFACS2cellexalvr}{cellexalvrR}(cellexalObj, facs.data)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{facs.data}{A matrix of surface marker intensities}
}
\description{
Adds FACS index to a cellexalvrObj
}
\details{
This function adds a table object into the index slot of the cellexalvrR obejct.
}
