% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMDS2cellexalvr.R
\docType{methods}
\name{addMDS2cellexalvr}
\alias{addMDS2cellexalvr}
\alias{addMDS2cellexalvr,cellexalvr-method}
\title{description of function addMDS2cellexalvr}
\arguments{
\item{name}{A name for the object (default = graph<n>)}

\item{cellexalObj}{the cellexalvr object}

\item{mdsmatrix}{the mds 3D matrix}

\item{name}{the name of this MDS e.g. 'PCA' default=NULL}
}
\description{
Adds an externally created MDS to this object
}
