% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportFunctions.R, R/addElements.R
\name{addMDS2cellexalvr}
\alias{addMDS2cellexalvr}
\alias{addMDS2cellexalvr}
\title{Adds mds coordinates to a cellexalvrObj}
\usage{
addMDS2cellexalvr(cellexalObj, mdsmatrix, name = NULL)

addMDS2cellexalvr(cellexalObj, mdsmatrix, name = NULL)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{mdsmatrix}{A matrix of coordinates}

\item{name}{A name for the object (default = graph<n>)}

\item{cellexalObj}{A cellexalvr object}

\item{mdsmatrix}{A matrix of coordinates}

\item{name}{A name for the object (default = graph<n>)}
}
\description{
Adds mds coordinates to a cellexalvrObj

Adds mds coordinates to a cellexalvrObj
}
