% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{addSelection}
\alias{addSelection}
\alias{addSelection,cellexalLinear,cellexalvrR-method}
\title{savely add a cellexalLinear object to a cellexalvR object}
\usage{
addSelection(x, cellexalObj, upstreamSelection = NULL)

\S4method{addSelection}{cellexalLinear,cellexalvrR}(x, cellexalObj, upstreamSelection = NULL)
}
\arguments{
\item{x}{the cellexalLinear object}

\item{cellexalObj}{the cellexalvrR object to add the linear selection as group}

\item{upstreamSelection}{which group was the basis for this linearSelection?}
}
\description{
add this cellexalLinear to the corresponding cellexal object
}
\details{
Method to copy a selection from cellexalObj to cellexalObj.
This function makes sure that the linear grouping is fitting into the object.
}
