% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchPointAnalysis.R
\name{branch.point.analysis}
\alias{branch.point.analysis}
\title{Performs a branch-point analysis identifying genes that have opposite changes in two branches stemming from one trunk in a pseudotime analysis}
\usage{
branch.point.analysis(cellexalObj, cellidfile, maxsig, outfile)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{maxsig}{The number of differentials to be returned}

\item{outfile}{The name of the output file}
}
\description{
Performs a branch-point analysis identifying genes that have opposite changes in two branches stemming from one trunk in a pseudotime analysis
}
\keyword{analysis}
\keyword{branch}
\keyword{point}
