% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalvrClass.R
\docType{class}
\name{cellexalGrouping-class}
\alias{cellexalGrouping-class}
\title{cellexalGrouping class definition}
\description{
An object to structure multi group selections
}
\details{
The cellexalGrouping class represents exactly one multi group selection in VR.
}
\section{Slots}{

\describe{
\item{\code{gname}}{the group name}

\item{\code{selectionFile}}{the VR selection file that is the basis for this grouping}

\item{\code{grouping}}{the numeric group ids for the R representation}

\item{\code{VRgrouping}}{the numeric group ids for the VR process (differs from the R)}

\item{\code{order}}{the order the cells have been selected in the VR process}

\item{\code{drc}}{the drc name this object has been selected from}

\item{\code{col}}{the color vector for these groups}

\item{\code{error}}{a string vector that contains all error messages}

\item{\code{timeObj}}{an optional slot to store a cellexalTime object}

\item{\code{heatmapBasename}}{the filename basis for the heatmap related to this grouping}
}}

