% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalvrClass.R
\docType{class}
\name{cellexalTime-class}
\alias{cellexalTime-class}
\title{cellexalTime class definition}
\description{
A simple wrapper to handle the linear order and respective color mappings.
}
\details{
the cellexalTime class is representing one linear ordered cell group
}
\section{Slots}{

\describe{
\item{\code{dat}}{all (drc) data needed for plotting and group creation}

\item{\code{gname}}{the group name}

\item{\code{drc}}{the drc name this object has been selected from}

\item{\code{error}}{the error message if a catched not fatal error has occured}

\item{\code{geneClusters}}{a list of gene clusters that are linked to a timeline}

\item{\code{id}}{the md5sum representation of thei object's dat (to prohibit duplicates)}

\item{\code{parentSelection}}{the cellexalGrouping name that was basis for this linear order}
}}

