% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrClass.R
\docType{class}
\name{cellexalvr}
\alias{cellexalvr}
\title{cellexalvr}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
cellexalvr
}
\description{
An R S4 class that represents the backend for the cellexal VR application
}
\details{
Class defintion of cellexalvr
}
\section{Fields}{

\describe{
\item{\code{data}}{a matrix storing the expression values}

\item{\code{meta.cell}}{a matrix storing the sample meta.gene data (data column)}

\item{\code{meta.gene}}{a matrix storing the gene meta.gene data (data row)}

\item{\code{mds}}{list of mds representations of the whole data frame}

\item{\code{index}}{a matrix of experimental sample descriptions e.g. FACS levels}

\item{\code{tfs}}{a vector of transcription factor names?}

\item{\code{stats}}{a list to harbor stat data tables. 
The stat tables need to have the same gene order as the data table}

\item{\code{name}}{the name of the object (mainly unused)}

\item{\code{usedObj}}{a list to store temporary data like color information and other loosely linked data}
}}

\keyword{datasets}
