% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeIdent.R
\docType{methods}
\name{changeIdent}
\alias{changeIdent}
\alias{changeIdent,cellexalvr-method}
\title{description of function changeIdent}
\arguments{
\item{seuratObj}{The seurat object to be changed}

\item{new.idents}{The new cell names}
}
\description{
Adds new cell identities to the seurat object
}
