% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\docType{methods}
\name{check}
\alias{check}
\alias{check,cellexalvrR-method}
\title{cellexal internal checks}
\usage{
check(x)
}
\arguments{
\item{x}{the cellexalvrR object}
}
\description{
checks cellexalvrR internals
}
\details{
check  cellexalvrR sanity check.
}
