% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{checkTime}
\alias{checkTime}
\alias{checkTime,cellexalTime-method}
\title{description of function checkTime}
\usage{
checkTime(x, cellexalObj)

\S4method{checkTime}{cellexalTime,cellexalvrR}(x, cellexalObj)
}
\arguments{
\item{x}{the cellexalTime object}

\item{cellexalObj}{an optional cellexalvrR object - if given ONLY the drc models are checked.}
}
\description{
checks for NA elements in the table and removes them

checks for NA elements in the table and removes them
}
\details{
The sanity check for a linear grouping.
If problems are detected the problems will be fixed.

When presented with a cellexalObj this function makes sure the object fits to the cellexalObj.
}
