% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{checkTime}
\alias{checkTime}
\alias{checkTime,cellexalTime-method}
\title{description of function check}
\arguments{
\item{x}{the cellexalTime object}

\item{cellexalObj}{an optional cellexalvrR object - if given ONLY the drc models are checked.}
}
\description{
checks for NA elements in the table and removes them
}
