% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrExportFunctions.R
\name{checkVRfiles}
\alias{checkVRfiles}
\title{checkVRfiles: Checks the existance of all VR specific files and re-runs the export function if any is missing.}
\usage{
checkVRfiles(cellexalvr, path)
}
\arguments{
\item{path}{the outpath to check}

\item{cellexalObj}{the cellexal object}
}
\description{
checkVRfiles: Checks the existance of all VR specific files and re-runs the export function if any is missing.
}
