% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlotHeatmaps.R
\docType{methods}
\name{clusterGenes}
\alias{clusterGenes}
\alias{clusterGenes,cellexalLinear-method}
\alias{clusterGenes,matrix-method}
\title{cluster the genes for a linearSelection analysis}
\usage{
clusterGenes(
  x,
  deg.genes = NULL,
  info = NULL,
  cellexalObj = NULL,
  geneclusters = NULL
)

\S4method{clusterGenes}{cellexalLinear}(
  x,
  deg.genes = NULL,
  info = NULL,
  cellexalObj = NULL,
  geneclusters = NULL
)

\S4method{clusterGenes}{matrix}(
  x,
  deg.genes = NULL,
  info = NULL,
  cellexalObj = NULL,
  geneclusters = NULL
)
}
\arguments{
\item{x}{either the z-scored matrix or a cellexalLinear object}

\item{deg.genes}{a list of genes}

\item{info}{the group to cluster the genes for (list)}

\item{cellexalObj}{if x is a cellexalLinear object this is necessary to create the zscored matrix.}

\item{geneclusters}{ovverride the WSS based optimal group count search (default NULL)}
}
\description{
get a - hopefully - optimal grouing of a list of genes
}
\details{
Tries to answer the question: how do these genes differ over the linearSelection.
It answers this in a graphical, not a statistical way.
Hence you can feed whichever genelist you like into this function.
}
