% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{collapseTime}
\alias{collapseTime}
\alias{collapseTime,cellexalTime-method}
\title{sum up cells based on the pseudo time}
\usage{
collapseTime(x, to = 1000)

\S4method{collapseTime}{cellexalTime}(x, to = 1000)
}
\arguments{
\item{x}{the cellexalTime object}

\item{to}{the max amnount of time points}
}
\description{
return the color in exactly the right format to color the names
}
\details{
collapse the time to a max of 1000 enrties rendering the timeline incompatible with cellexalvrR
but allowing the correlation of collapsed expression to the time.
}
