% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{color}
\alias{color}
\alias{color,cellexalTime-method}
\title{return a pseudo time color vector for a given cell names list}
\usage{
color(x, names)

\S4method{color}{cellexalTime}(x, names)
}
\arguments{
\item{x}{the cellexalTime object}

\item{names}{the cells to color with the time color}
}
\description{
return the color in exactly the right format to color the given cell names
}
\details{
returns a color vector based on the intern col column and the cell names
Cells that are not part of this selection will get gray(0.6) as color.
}
