% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\name{compactTime,cellexalvrR-method}
\alias{compactTime,cellexalvrR-method}
\title{creates compacted expression data for all timelines using FastWilcoxTest::collapse()}
\usage{
\S4method{compactTime}{cellexalvrR}(cellexalObj)
}
\description{
creates compacted expression data for all timelines using FastWilcoxTest::collapse()
}
