% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{compactTime}
\alias{compactTime}
\alias{compactTime,cellexalvrR-method}
\title{merge cells based on time order}
\usage{
compactTime(cellexalObj)

\S4method{compactTime}{cellexalvrR}(cellexalObj)
}
\arguments{
\item{cellexalObj}{if x is a cellexalTime object this is necessary to create the zscored matrix.}
}
\description{
create a global gene z-score and merged virtual samples over the timeline
}
