% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{compactTimeZscore}
\alias{compactTimeZscore}
\alias{compactTimeZscore,cellexalTime-method}
\title{description of function plot}
\usage{
compactTimeZscore(x, deg.genes, info, cellexalObj)
}
\arguments{
\item{x}{either the z-scored matrix or a cellexalTime object}

\item{deg.genes}{a list of genes  cluster the genes for (list)}

\item{cellexalObj}{if x is a cellexalTime object this is necessary to create the zscored matrix.}
}
\description{
create a global gene z-score and virtual samples over the timeline
}
