% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{compareGeneClusters}
\alias{compareGeneClusters}
\alias{compareGeneClusters,cellexalTime-method}
\title{description of function color}
\arguments{
\item{x}{the cellexalTime object}

\item{other}{the other timeline (the parentSelection has to match)}

\item{cellexalObj}{the cellexalObj to add the report to.}

\item{altGroupNames}{you can give alternative group names here}

\item{color}{a vector of colors - one for each group}

\item{GOIs}{a different group io genes - might need to re-run the analysis}
}
\description{
return the color in exactly the right format to color the names
}
\details{
Compare gene clusters between different (reported) timelines.
The main usage for this function is to run the same genes on the same timeline,
but with different samples like young and old, healthy and sick or teated and untreated.
This function only allows to compare two gene groupings.

The report will for each gene cluster show the heatmap here and the heatmap in the other timeline
for each overlapping gene list.
}
