% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{createReport}
\alias{createReport}
\alias{createReport,cellexalTime-method}
\title{description of function plot}
\arguments{
\item{x}{the object}

\item{cellexalObj}{the object to get the data from}

\item{info}{a cellexalvrR grouping info list}

\item{deg.genes}{a list of genes to create the report for}
}
\description{
3D plot the time with a line in the time
}
\details{
This report is the main go to analysis for a timeline.
It will produce gene clusters and plot the mean expression of these genes clusters as beautiful xy plots.
It also populated the CellexalTime geneClusters slot with a list entry containing
the gene clusters (clusters), the max 1000 cells zscored data (matrix) and the location of the main figure (png file path; figure)

This enables the cellexalTime to be compared to another cellexalTime object using the function compareGeneClusters.
compareGeneClusters should best be run after a createDetailedComparison section has been added?!
}
