% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{createReport}
\alias{createReport}
\alias{createReport,character,cellexalvrR,character-method}
\alias{createReport,cellexalLinear,cellexalvrR,character-method}
\alias{createReport,cellexalLinear,cellexalvrR,cellexalLinear-method}
\alias{createReport,cellexalLinear,cellexalvrR,cellexalGrouping-method}
\title{create a report for one cellexalLinear object (interactive mode only)}
\usage{
createReport(x, cellexalObj, info, deg.genes = NULL, num.sig = 250)

\S4method{createReport}{character,cellexalvrR,character}(x, cellexalObj, info, deg.genes = NULL, num.sig = 250)

\S4method{createReport}{cellexalLinear,cellexalvrR,character}(x, cellexalObj, info, deg.genes = NULL, num.sig = 250)

\S4method{createReport}{cellexalLinear,cellexalvrR,cellexalLinear}(x, cellexalObj, info, deg.genes = NULL, num.sig = 250)

\S4method{createReport}{cellexalLinear,cellexalvrR,cellexalGrouping}(x, cellexalObj, info, deg.genes = NULL, num.sig = 250)
}
\arguments{
\item{x}{the cellexalLinear object or a linear grouping name}

\item{cellexalObj}{the object to get the data from}

\item{info}{a cellexalGrouping or grouping name}

\item{deg.genes}{a list of genes to create the report for}

\item{num.sig}{number of signififcant genes (default 250)}
}
\description{
3D plot using linear order colors
}
\details{
This report is the main go to analysis for a linearSelection.
It will produce gene clusters and plot the mean expression of these genes clusters as beautiful xy plots.
It also populates the CellexalTime geneClusters slot with a list entries containing
the gene clusters (clusters), the max 1000 cells zscored data (matrix) and the location of the main figure (png file path; figure)

After this function two cellexalLinear objects can be compared to another using the function compareGeneClusters.
}
