% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{createTime}
\alias{createTime}
\alias{createTime,cellexalTime-method}
\title{craete the pseudo time for the selection}
\usage{
createTime(x, parentSelection = NULL)

\S4method{createTime}{cellexalTime}(x, parentSelection = NULL)
}
\arguments{
\item{x}{the object}

\item{parentSelection}{the selection name this time bases on.}
}
\description{
calculate time based on the internal table a b and c columns
}
\details{
This function utilizes slingshot to identify the longest possible pseudo 
timeline for this selection.
In short the drc model of the selected genes is clustered using kmeans
and the first and last selected cell is used to determine the start and end points of the
anlysis in the drc model.
Subsequently slignshot is used to calculate the pseudo time over the selection.
This function will create the same timeline if the same cells are used as input.
}
