% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineGOIs.R
\name{defineGOIs}
\alias{defineGOIs}
\title{Allows the user to define (G)enes (O)f (I)nterest lists in the object}
\usage{
defineGOIs(cellexalObj, name, genes, lables = NULL, ...)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{name}{the name of the GIO list (eg TFs or epigenetic)}

\item{genes}{a list of gene symbols that match to the @data rownames}

\item{lables}{a list of lables for the GIO column (default NULL)}
}
\description{
Allows the user to define (G)enes (O)f (I)nterest lists in the object
}
