% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineGOIs.R
\docType{methods}
\name{defineGOIs}
\alias{defineGOIs}
\alias{defineGOIs,cellexalvrR-method}
\title{Define special genes based on inbuilt data sets (TFs, ...)}
\usage{
defineGOIs(cellexalObj, name, genes, lables = NULL)

\S4method{defineGOIs}{cellexalvrR}(cellexalObj, name, genes, lables = NULL)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{name}{the name of the GIO list (eg TFs or epigenetic)}

\item{genes}{a list of gene symbols that match to the @data rownames}

\item{lables}{a list of lables for the GIO column (default NULL)}
}
\description{
Allows the user to define (G)enes (O)f (I)nterest lists in the object
}
\details{
CellexalvrR includes several tables describing gene functionality.
Namely TFs, epigenetic, CellCycle and CellSurface genes.
This function will match the CellexalvrR object against these lists and add a
column into the meta.gene table containing 0 and 1 values.

This function is used for the correlation networks in VR.
}
