% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Epigenetic}
\alias{Epigenetic}
\title{A simple table containing the data from 'http://epifactors.autosome.ru/' as from 21st September 2017}
\format{
data.frame
}
\usage{
Epigenetic
}
\description{
This table can be used to create the epigenetics DRC objects.
}
\keyword{data}
