% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportFunctions.R
\docType{methods}
\name{export2cellexalvr}
\alias{export2cellexalvr}
\alias{export2cellexalvr,cellexalvrR-method}
\title{description of function export2cellexalvr}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{forceDB}{re-write the db even if it exisis (default =F)}
}
\description{
Creates the base files needed to run the VR environment
}
