% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2cellexalvr.R
\docType{methods}
\name{export2cellexalvr}
\alias{export2cellexalvr}
\alias{export2cellexalvr,cellexalvr-method}
\title{description of function export2cellexalvr}
\usage{
export2cellexalvr(cellexalObj, path, forceDB = F)
}
\arguments{
\item{cellexalObj}{the cellexalvr object to be exported}

\item{path}{the path to create the files in}

\item{forceDB}{re-write the db even if it exisis (default =F)}

\item{forceDB}{re-create the expression data sglite database (default=FALSE)}
}
\description{
Creates all files necessary for the VR application.
}
