% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportFunctions.R
\docType{methods}
\name{export2cellexalvr}
\alias{export2cellexalvr}
\alias{export2cellexalvr,cellexalvrR-method}
\title{create the VR data folder necessary for CellexalVR}
\usage{
\S4method{export2cellexalvr}{cellexalvrR}(cellexalObj, path, forceDB = F,
  VRpath = NULL)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{path}{the oputpath to store the data in}

\item{forceDB}{re-write the db even if it exisis (default =F)}

\item{VRpath}{in order to re-color the data in the VR process the 
VR process needs the grouping names which will be exported if we get the correct path here}
}
\description{
Creates the base files needed to run the VR environment
}
\details{
This function creates all files necessary 
for the CellexalVR application to show this data.

Consult the CellexalVR documentation for further path requrements.
}
\examples{
dir.create ('data')
export2cellexalvr(cellexalObj, path='data') #function definition in file 'ExportFunctions.R'
}
