% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrExportFunctions.R
\name{export2cellexalvr}
\alias{export2cellexalvr}
\title{Creates the base files needed to run the VR environment}
\usage{
export2cellexalvr(cellexalObj)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}
}
\description{
Creates the base files needed to run the VR environment
}
