% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportFunctions.R
\docType{methods}
\name{export2cellexalvr}
\alias{export2cellexalvr}
\alias{export2cellexalvr,cellexalvrR-method}
\title{create the VR data folder necessary for CellexalVR}
\usage{
export2cellexalvr(x, path, forceDB = FALSE)

\S4method{export2cellexalvr}{cellexalvrR}(x, path, forceDB = FALSE)
}
\arguments{
\item{x}{A cellexalvr object}

\item{path}{the oputpath to store the data in}

\item{forceDB}{re-write the db even if it exisis (default =F)}
}
\description{
Creates the base files needed to run the VR environment
}
\details{
This function creates all files necessary 
for the CellexalVR application to show this data.

The folder with this data needs to resude in the CellexalVR\\DATA folder.
Consult the CellexalVR documentation for further path requrements.
}
\examples{
\dontrun{
dir.create ('data')
export2cellexalvr(x, path='data') #function definition in file 'ExportFunctions.R'
}
}
