% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{exportSelection}
\alias{exportSelection}
\alias{exportSelection,cellexalTime-method}
\title{description of function exportSelection}
\usage{
exportSelection(x, fname)
}
\arguments{
\item{x}{the object}

\item{fname}{the file to write the info to}
}
\description{
Export this time as selection file
}
\details{
This creates a selection.time file for the VR process.
}
