% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{exportSelection}
\alias{exportSelection}
\alias{exportSelection,cellexalTime-method}
\title{description of function exportSelection}
\arguments{
\item{x}{the object}

\item{fname}{the file to write the info to}
}
\description{
Export this time as selection file
}
