% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{exportSelection}
\alias{exportSelection}
\alias{exportSelection,cellexalTime-method}
\title{export a cellexalTime object for CellexalVR}
\usage{
exportSelection(x, fname)

\S4method{exportSelection}{cellexalTime}(x, fname)
}
\arguments{
\item{x}{the object}

\item{fname}{the file to write the info to}
}
\description{
Export this time as selection file
}
