% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUserGroups4vr.R
\docType{methods}
\name{exportUserGroups4vr}
\alias{exportUserGroups4vr}
\alias{exportUserGroups4vr,cellexalvrR-method}
\title{VR helper function exportUserGroups4vr}
\usage{
\S4method{exportUserGroups4vr}{cellexalvrR}(cellexalObj, path)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{path}{the outpath}
}
\description{
Creates a summary file for the vr process Creates a file groupings_info.txt in the

outfolder that contains the group name (in the R object) the numer of groups in the

selection and the number of cells in the whole group.
}
\details{
This is a 'VR' helper function that creates a dynamic file used to make the R groupings known to the VR process.

Users of the 'cellexalvrR' R class should not need this function.
}
\examples{
dir.create('data')
exportUserGroups4vr( cellexalObj, 'data') #function definition in file 'exportUserGroups4vr.R'
}
