% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\docType{methods}
\name{filename}
\alias{filename}
\alias{filename,cellexalvrR-method}
\title{description of function filename}
\usage{
filename(str)
}
\arguments{
\item{str}{a string vector}
}
\description{
simple function to join by '.' and remove whitespace.
}
