% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_forceAbsoluteUniqueSample.R
\docType{methods}
\name{forceAbsoluteUniqueSample}
\alias{forceAbsoluteUniqueSample}
\alias{forceAbsoluteUniqueSample,character-method}
\title{description of function forceAbsoluteUniqueSample}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{separator}{'_' or anything you want to set the separator to}
}
\description{
force absolute unique names in a vector by adding _<amount of repeats> to each value

if there is more than one repeat opf the value @exportMethod
}
