% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cellexalvrR.R
\docType{methods}
\name{forceAbsoluteUniqueSample}
\alias{forceAbsoluteUniqueSample}
\alias{forceAbsoluteUniqueSample,cellexalvrR-method}
\title{make the cell names unique}
\usage{
forceAbsoluteUniqueSample(x, separator = "_")

\S4method{forceAbsoluteUniqueSample}{cellexalvrR}(x, separator = "_")
}
\arguments{
\item{x}{the string vector you want to force into uniques}

\item{separator}{the separator between orig str and id ( default '_')}
}
\description{
This function adds _<id> to all duplicate values thereby enforcing uniques.
}
