% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\docType{methods}
\name{get.genes.cor.to}
\alias{get.genes.cor.to}
\alias{get.genes.cor.to,cellexalvrR-method}
\title{description of function get.genes.cor.to}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{gname}{The required gene}

\item{output}{the outfile}

\item{is.smarker}{Whether the supplied gene is a surface marker (default =F)}

\item{gname}{TEXT MISSING}

\item{output}{TEXT MISSING}

\item{is.smarker}{TEXT MISSING default=F}
}
\description{
Gets positively and negatively correlated genes to a chosen gene
}
\keyword{correlation}

