% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\name{get.genes.cor.to}
\alias{get.genes.cor.to}
\title{Gets positively and negatively correlated genes to a chosen gene}
\usage{
get.genes.cor.to(cellexalObj, gname, output, is.smarker = F)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{gname}{The required gene}

\item{output}{the outfile}

\item{is.smarker}{Whether the supplied gene is a surface marker (default =F)}
}
\description{
Gets positively and negatively correlated genes to a chosen gene
}
\keyword{correlation}
