% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\docType{methods}
\name{get.genes.cor.to,character-method}
\alias{get.genes.cor.to,character-method}
\title{description of function get.genes.cor.to (character) #function definition in file 'cellexalvrFunctions.R'}
\usage{
\S4method{get.genes.cor.to}{character}(cellexalObj, gname, output = NULL,
  is.smarker = F, cpp = T)
}
\arguments{
\item{cellexalObj}{the cellexalObj.RData file}

\item{gname}{The required gene}

\item{output}{the outfile}

\item{is.smarker}{Whether the supplied gene is a surface marker (default =F)}

\item{cpp}{use the c++ cor implementation (default = TRUE)}
}
\description{
preload the cellexalObj
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: cellexalvrR
}}

