% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentials.R
\docType{methods}
\name{getDifferentials}
\alias{getDifferentials}
\alias{getDifferentials,cellexalvrR-method}
\title{description of function getDifferentials}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{deg.method}{The method to use to find DEGs}

\item{numsig}{The number of differentials to be returned}

\item{deg.method}{TEXT MISSING default=c("anova"}

\item{"edgeR"}{TEXT MISSING default=c("anova"}

\item{"MAST")}{TEXT MISSING default=c("anova"}

\item{num.sig}{TEXT MISSING}
}
\description{
Creates a heatmap from a selection of groups
}
\keyword{DEGs}
