% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentials.R
\docType{methods}
\name{getDifferentials}
\alias{getDifferentials}
\alias{getDifferentials,cellexalvrR-method}
\title{VR helper function getDifferentials}
\usage{
\S4method{getDifferentials}{cellexalvrR}(
  x,
  cellidfile,
  deg.method = c("wilcox", "Seurat_wilcox", "bimod", "roc", "t", "tobit", "poisson",
    "negbinom", "MAST", "DESeq2", "anova"),
  num.sig = 250,
  Log = TRUE,
  logfc.threshold = 0.1,
  minPct = 0.1,
  onlyPos = TRUE
)
}
\arguments{
\item{x, }{cellexalvrR object}

\item{cellidfile}{file containing cell IDs}

\item{deg.method}{The method to use to find DEGs ( 'wilcox', 'Seurat wilcox', 'bimod', 'roc', 't', 'tobit', 'poisson', 'negbinom', 'MAST', 'DESeq2' )}

\item{num.sig}{number of differnetial genes to return (250)}

\item{Log}{log the results (default=TRUE)}

\item{logfc.threshold}{the Seurat logfc.threshold option (default here 1 vs 0.25 in Seurat)}

\item{minPct}{the minium percent expressing cells in a group (default 0.1)}

\item{onlyPos}{select only genes showing an higher expression in the group (default =T)}
}
\value{
the cellexalvrr object with the stats table stored in x@usedObj$sigGeneLists$Cpp[[x@usedObj$lastGroup]]
and significant genes can be accessed in the x@usedObj$deg.genes slot.
}
\description{
Creates a heatmap from a selection of groups
The Seurat based statsictsics is applied only to genes expressed in at least 1 percent of the cells.
}
\details{
Identify differentially expressed genes.

This function makes three statistics available for the VR process 
(1) 'timeline' will automaticly be choosen it there is only one group in the data
(2) 'wilcox' a c++ re-implementation of the Seurat::FindAllMarkers function (default)
(3) 'Seurat_wilcox' the original Seurat::FindAllMarkers function (~10x slower than the c++ version and currently not working)
}
\examples{
\dontrun{
getDifferentials( x,  cellidfile= 'User.group.2', deg.method='wilcox')@usedObj$deg.genes #function definition in file 'getDifferentials.R'
}
}
\keyword{DEGs}
