% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentials.R
\docType{methods}
\name{getDifferentials}
\alias{getDifferentials}
\alias{getDifferentials,cellexalvrR-method}
\title{VR helper function getDifferentials}
\usage{
\S4method{getDifferentials}{cellexalvrR}(cellexalObj, cellidfile,
  deg.method = c("wilcox", "Seurat_wilcox", "bimod", "roc", "t", "tobit",
  "poisson", "negbinom", "MAST", "DESeq2", "anova"), num.sig = 250,
  Log = TRUE, logfc.threshold = 1, minPct = 0.1, onlyPos = TRUE)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{deg.method}{The method to use to find DEGs ( 'wilcox', 'Seurat wilcox', 'bimod', 'roc', 't', 'tobit', 'poisson', 'negbinom', 'MAST', 'DESeq2' )}

\item{num.sig}{number of differnetial genes to return (250)}

\item{Log}{log the results (default=TRUE)}

\item{logfc.threshold}{the Seurat logfc.threshold option (default here 1 vs 0.25 in Seurat)}

\item{minPct}{the minium percent expressing cells in a group (default 0.1)}

\item{onlyPos}{select only genes showing an higher expression in the group (default =T)}
}
\value{
a list of the top differential genes (num.sig)
}
\description{
Creates a heatmap from a selection of groups
The Seurat based statsictsics is applied only to genes expressed in at least 1 percent of the cells.
}
\details{
Identify differentially expressed genes.

This function makes three statistics available for the VR process 
(1) 'linear' correlation for one group selections
(2) 'wilcox' a c++ re-implementation of the Seurat::FindAllMarkers function
(3) 'Seurat_wilcox' the original Seurat::FindAllMarkers function (~10x slower than the c++ version)
}
\examples{
\dontrun{
getDifferentials( cellexalObj,  cellidfile= 'User.group.2', deg.method='wilcox') #function definition in file 'getDifferentials.R'
}
}
\keyword{DEGs}
