% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentials.R
\docType{methods}
\name{getDifferentials}
\alias{getDifferentials}
\alias{getDifferentials,cellexalvrR-method}
\alias{getDifferentials,character-method}
\title{main Statistics function for cellexalvrR}
\usage{
getDifferentials(
  x,
  cellidfile,
  deg.method = "wilcox",
  num.sig = 250,
  Log = TRUE,
  logfc.threshold = 1,
  minPct = 0.1,
  onlyPos = TRUE,
  report4genes = NULL
)

\S4method{getDifferentials}{cellexalvrR}(
  x,
  cellidfile,
  deg.method = "wilcox",
  num.sig = 250,
  Log = TRUE,
  logfc.threshold = 0.1,
  minPct = 0.1,
  onlyPos = TRUE,
  report4genes = NULL
)

\S4method{getDifferentials}{character}(
  x,
  cellidfile,
  deg.method = "wilcox",
  num.sig = 250,
  Log = TRUE,
  logfc.threshold = 1,
  minPct = 0.1
)
}
\arguments{
\item{x, }{cellexalvrR object}

\item{cellidfile}{file containing cell IDs}

\item{deg.method}{The method to use to find DEGs (only 'wilcox' supported at the moment)}

\item{num.sig}{number of differnetial genes to return (250)}

\item{Log}{log the results (default=TRUE)}

\item{logfc.threshold}{the Seurat logfc.threshold option (default here 1 vs 0.25 in Seurat)}

\item{minPct}{the minium percent expressing cells in a group (default 0.1)}

\item{onlyPos}{select only genes showing an higher expression in the group (default =T)}

\item{report4genes}{a list of genes you want to get a report on.}
}
\value{
the cellexalvrr object with the stats table stored in x@usedObj$sigGeneLists$Cpp[[x@usedObj$lastGroup]]
and significant genes can be accessed in the x@usedObj$deg.genes slot.
}
\description{
Calculates the statistics for one CellexalVR selection file.
}
\details{
Identify differentially expressed genes.

This function makes two statistics available for the VR process 
(1) 'timeline' will automaticly be choosen if there is only one group in the data
(2) 'wilcox' a c++ re-implementation of the Seurat::FindAllMarkers function (default)
}
\examples{
\dontrun{
getDifferentials( x,  cellidfile= 'User.group.2', deg.method='wilcox')@usedObj$deg.genes 
}
}
\keyword{DEGs}
