% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentials.R
\docType{methods}
\name{getDifferentials,character-method}
\alias{getDifferentials,character-method}
\title{description of function getDifferentials}
\usage{
\S4method{getDifferentials}{character}(
  x,
  cellidfile,
  deg.method = c("wilcox", "Seurat_wilcox", "bimod", "roc", "t", "tobit", "poisson",
    "negbinom", "MAST", "DESeq2", "anova"),
  num.sig = 250,
  Log = TRUE,
  logfc.threshold = 1,
  minPct = 0.1
)
}
\arguments{
\item{x}{the cellexalvrR.RData file}

\item{cellidfile}{file containing cell IDs}

\item{deg.method}{The method to use to find DEGs ( 'wilcox', 'Seurat wilcox', 'bimod',
'roc', 't', 'tobit', 'poisson', 'negbinom', 'MAST', 'DESeq2')}

\item{num.sig}{number of differnetial genes to return (250)}

\item{Log}{log the results (default=TRUE)}

\item{logfc.threshold}{the Seurat logfc.threshold option (default here 1 vs 0.25 in Seurat)}

\item{minPct}{the minium percent expressing cells in a group (default 0.1)}
}
\description{
preload the cellexalObj
}
\section{Functions}{
\itemize{
\item \code{getDifferentials,character-method}: cellexalvrR
}}

\keyword{DEGs}
