% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session2HTMLtable.R
\docType{methods}
\name{group2HTMLtable}
\alias{group2HTMLtable}
\alias{group2HTMLtable,cellexalvrR-method}
\title{create a log table from a grouping}
\arguments{
\item{x, }{cellexalvrR object}

\item{cellidfile}{file containing cell IDs}

\item{names}{optional vector of names for the groupings (grouping info is numeric)}
}
\description{
create an HTML table from a user grouping
}
\details{
Convert a session into a html table that can be added to the log
}
