% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupingInfo.R
\docType{methods}
\name{groupingInfo}
\alias{groupingInfo}
\alias{groupingInfo,cellexalvr-method}
\title{description of function groupingInfo}
\usage{
groupingInfo(cellexalObj, gname = NULL)
}
\arguments{
\item{cellexalObj}{the cellexalvr object}

\item{gname}{the group name you get the data for ( default last added group)}
}
\description{
returns the information stored for the last grouping read
or the group defined in the gname option
}
