% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupingInfo.R
\docType{methods}
\name{groupingInfo}
\alias{groupingInfo}
\alias{groupingInfo,cellexalvrR-method}
\title{get information on a sample grouping}
\usage{
groupingInfo(cellexalObj, gname = NULL)

\S4method{groupingInfo}{cellexalvrR}(cellexalObj, gname = NULL)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{gname}{The optional group name to get info on a specific grouping (not the last)}
}
\value{
a cellexalGrouping object
}
\description{
returns the information stored for the last grouping read
}
\details{
groupingInfo collects all information about one grouping from 
the cellexalvrR internals and returns them as a cellealGrouping object.
}
