% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R, R/03cellexalGrouping.R
\docType{methods}
\name{HTMLtable}
\alias{HTMLtable}
\alias{HTMLtable,cellexalTime-method}
\alias{HTMLtable,cellexalGrouping-method}
\title{description of function HTMLtable}
\usage{
HTMLtable(x)

\S4method{HTMLtable}{cellexalTime}(x)

\S4method{HTMLtable}{cellexalGrouping}(x)
}
\arguments{
\item{x}{the cellexalTime object}
}
\description{
return the HTMLtable in exactly the right format
}
\details{
A log specific function returning a html formate table describing the timeline
}
