% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{HTMLtable}
\alias{HTMLtable}
\alias{HTMLtable,cellexalTime-method}
\title{description of function HTMLtable}
\arguments{
\item{x}{the cellexalTime object}
}
\description{
return the HTMLtable in exactly the right format
}
\details{
returns a html formate table describing one timeline
}
