% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R, R/03cellexalGrouping.R
\docType{methods}
\name{HTMLtable}
\alias{HTMLtable}
\alias{HTMLtable,cellexalTime-method}
\alias{HTMLtable,cellexalGrouping-method}
\title{description of function HTMLtable}
\usage{
HTMLtable(x)

\S4method{HTMLtable}{cellexalGrouping}(x)
}
\arguments{
\item{x}{the cellexalGrouping}
}
\description{
return the HTMLtable in exactly the right format

convert a H5 annotation (any name) table to a data table
}
\details{
returns a html formate table describing one timeline

Create an HTML table from a cellexalGrouping.
These tables are part of the log html's and show group color as well as cell count, VR- and R-ID.
}
