% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{invert}
\alias{invert}
\alias{invert,cellexalLinear-method}
\title{reverse the time order for a cellexalLinear object}
\usage{
invert(x)

\S4method{invert}{cellexalLinear}(x)
}
\arguments{
\item{x}{the object}
}
\description{
reverse the time orientaion
}
\details{
inverts the time in the cellexalLinear object.

Occational the algorithm fails and inverts the linear order.
This function is used to easily invert a (inverted) linearSelection that 
runs from end to start instead the otherway around.
}
