% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\docType{methods}
\name{loadObject}
\alias{loadObject}
\alias{loadObject,cellexalvrR-method}
\title{description of function loadObject}
\arguments{
\item{fname}{the file to load or a cellexalvr object}

\item{fname}{TEXT MISSING}

\item{maxwait}{TEXT MISSING default=50}
}
\description{
Loads the cellexalvr object, if the fname is a file
}
\keyword{load}

