% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockedSave.R
\docType{methods}
\name{loadObject}
\alias{loadObject}
\alias{loadObject,character-method}
\title{description of function loadObject}
\usage{
\S4method{loadObject}{character}(fname, maxwait = 50)
}
\arguments{
\item{fname}{the file to load or a cellexalvr object}

\item{maxwait}{stop after maxwait seconds default=50}
}
\description{
Loads the cellexalvr object, if the fname is a file
}
\details{
loadObject has thread functionallity looking for a lock file and waiting for 'maxwait' seconds 
before reporting a failed attempt.
}
\keyword{load}
