% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\name{loadObject}
\alias{loadObject}
\title{Loads the cellexalvr object, if the fname is a file}
\usage{
loadObject(fname, maxwait = 50)
}
\arguments{
\item{fname}{the file to load or a cellexalvr object}
}
\description{
Loads the cellexalvr object, if the fname is a file
}
\keyword{load}
