% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockedSave.R
\docType{methods}
\name{lockedSave}
\alias{lockedSave}
\alias{lockedSave,cellexalvrR-method}
\title{description of function lockedSave}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{path}{the output path}

\item{path}{TEXT MISSING}
}
\description{
Saving the RData in the VR tool might create a problem. Hence this function will

save the cellexalObj in a controlled way.
}
\keyword{lockedSave}
