% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockedSave.R
\docType{methods}
\name{lockedSave}
\alias{lockedSave}
\alias{lockedSave,cellexalvrR-method}
\title{thread save save method for cellexalvrR objects}
\usage{
lockedSave(cellexalObj, path = NULL, what = NULL)

\S4method{lockedSave}{cellexalvrR}(cellexalObj, path = NULL)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{path}{the output path}

\item{what}{which part needs saving? (default NULL == all)}
}
\description{
Saving the RData in the VR tool might create a problem. Hence this function will

save the cellexalObj in a controlled way. Locked save removes all parts from the file system.
}
\details{
A thread save saving of the object.
}
\keyword{lockedSave}
