% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logHeatmap.R
\docType{methods}
\name{logHeatmap,character-method}
\alias{logHeatmap,character-method}
\title{description of function logHeatmap}
\usage{
\S4method{logHeatmap}{character}(cellexalObj, genes, png, grouping, ...)
}
\arguments{
\item{cellexalObj}{the cellexalvrR file}

\item{genes}{the genes displayed on the heatmap}

\item{png}{the VR generated heatmap (png)}

\item{grouping}{the grouping file used to create this heatmap}

\item{...}{options you want to send to the ontologyLogPage() function #function definition in file 'ontologyLogPage.R'}
}
\description{
create one Heatmap session report page
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: cellexalvrR
}}

