% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNetwork.R
\docType{methods}
\name{logNetwork,character-method}
\alias{logNetwork,character-method}
\title{description of function logNetwork}
\usage{
\S4method{logNetwork}{character}(cellexalObj, genes = NULL, png,
  grouping, ...)
}
\arguments{
\item{cellexalObj}{the cellexalObj.RData file}

\item{genes}{the genes displayed on the network}

\item{png}{the VR generated network (png)}

\item{grouping}{the grouping file used to create this network}

\item{...}{options you want to send to the ontologyLogPage() function #function definition in file 'ontologyLogPage.R'}
}
\description{
preload the cellexalObj.RData file
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: cellexalvrR
}}

