% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTimeLine.R
\name{logTimeLine,character-method}
\alias{logTimeLine,character-method}
\title{description of function logTimeLine}
\usage{
\S4method{logTimeLine}{character}(cellexalObj, stats, genes, info, png, timeInfo, GOIs = NULL)
}
\arguments{
\item{cellexalObj}{the cellexalvrR file}

\item{stats}{the correlation statistics}

\item{genes}{the genes to display on the heatmap}

\item{info}{the original grouping information list}

\item{timeInfo}{the time grouping information list}

\item{GOIs}{an optional vector of genes to plot rolling sum graphs for.}
}
\description{
create one Heatmap session report page
}
\section{Functions}{
\itemize{
\item \code{logTimeLine,character-method}: cellexalvrR
}}

