% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTimeLine.R
\docType{methods}
\name{logTimeLine}
\alias{logTimeLine}
\alias{logTimeLine,cellexalvrR-method}
\alias{logTimeLine,character-method}
\title{add the timeline information to the log system}
\usage{
logTimeLine(
  cellexalObj,
  stats,
  genes = NULL,
  info,
  png,
  timeInfo,
  GOIs = NULL,
  text = NULL
)

\S4method{logTimeLine}{cellexalvrR}(
  cellexalObj,
  stats,
  genes = NULL,
  info,
  png,
  timeInfo,
  GOIs = NULL,
  text = NULL
)

\S4method{logTimeLine}{character}(
  cellexalObj,
  stats,
  genes = NULL,
  info,
  png,
  timeInfo,
  GOIs = NULL,
  text = NULL
)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{stats}{the correlation statistics}

\item{genes}{the genes to display on the heatmap}

\item{info}{the original grouping information list}

\item{png}{the heatmap of the rolling sum data}

\item{timeInfo}{the time grouping information list}

\item{GOIs}{an optional vector of genes to plot rolling sum graphs for.}

\item{text}{additional text for the HTML file (default = NULL)}
}
\description{
preload the object before creating one Heatmap session report page
}
\details{
logTimeLine will create a section in the log document including 
(1) the DRC the grouping was selected from (colored 2D)
(2) the heatmap itself
(3) a GO analysis of the genes displayed in the heatmap (using ontologyLogPage()) #function definition in file 'ontologyLogPage.R'
}
