% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cell.meta.from.df.R
\docType{methods}
\name{make.cell.meta.from.df}
\alias{make.cell.meta.from.df}
\alias{make.cell.meta.from.df,data.frame-method}
\title{create a 0/1 table for a data.frame and a list of colnames}
\usage{
make.cell.meta.from.df(metad, rq.fields)

\S4method{make.cell.meta.from.df}{data.frame}(metad, rq.fields)
}
\arguments{
\item{metad}{A dataframe of per cell metadata}

\item{rq.fields}{A vector of name specifiying which columns should me made into metadata}
}
\description{
Creates a meta cell matrix from a supplied dataframe from required fields
}
\details{
The cell meta for cellexalvrR is a list of columns with 0/1 information for all groupings.
As this is no common format this function will convert a multi named grouping 
like CellType=c('HSC', MPP1', MPP2','HSC',...) into 3 columns that CellexalVR can use.
}
\keyword{cell}
\keyword{metadata}
