% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cellexalvr.heatmap.R
\docType{methods}
\name{make.cellexalvr.heatmap}
\alias{make.cellexalvr.heatmap}
\alias{make.cellexalvr.heatmap,cellexalvr-method}
\title{description of function make.cellexalvr.heatmap}
\usage{
make.cellexalvr.heatmap(cvrObj, cellidfile, num.sig, outfile)
}
\arguments{
\item{cvrObj}{the cellexavr object or the file to load the object from}

\item{cellidfile}{the grouping or a file to read a grouping from}

\item{num.sig}{the amount of significant genes to plot}

\item{outfile}{the file to plot to (png)}
}
\description{
A VR function that creates a heatmap
}
