% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringFunctions.R
\docType{methods}
\name{make.cellexalvr.heatmap.list}
\alias{make.cellexalvr.heatmap.list}
\alias{make.cellexalvr.heatmap.list,cellexalvrR-method}
\title{description of function make.cellexalvr.heatmap.list}
\arguments{
\item{cvrObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{num.sig}{The number of differentials to be returned}

\item{outfile}{The name of the output file}
}
\description{
Creates a list of genes from which a heatmap is made in Unity
}
\keyword{heatmap}

