% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cellexalvr.heatmap.list.R
\docType{methods}
\name{make.cellexalvr.heatmap.list}
\alias{make.cellexalvr.heatmap.list}
\alias{make.cellexalvr.heatmap.list,cellexalvrR-method}
\alias{make.cellexalvr.heatmap.list,character-method}
\title{Create the gene list files for VR}
\usage{
make.cellexalvr.heatmap.list(
  cvrObj,
  cellidfile,
  num.sig,
  outfile,
  stats_method = "wilcox"
)

\S4method{make.cellexalvr.heatmap.list}{cellexalvrR}(
  cvrObj,
  cellidfile,
  num.sig,
  outfile,
  stats_method = "wilcox"
)

\S4method{make.cellexalvr.heatmap.list}{character}(
  cvrObj,
  cellidfile,
  num.sig,
  outfile,
  stats_method = NA
)
}
\arguments{
\item{cvrObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs or the grouping name}

\item{num.sig}{The number of differentials to be returned}

\item{outfile}{The name of the output file}

\item{stats_method}{the stats method to use see getDifferentials()}
}
\description{
Creates an outfile for CellexalVR containing a list of genes from which a heatmap is made in VR.
Internally this function uses the getDifferentials() function and writes gene list to the outfile.
}
\details{
This is a function called from CellexalVR.

Its main function is to store the result from getDifferentials() in a file  #function definition in file 'getDifferentials.R'
to be read by CellexalVR.
}
\keyword{differential,}
\keyword{gene}
\keyword{list}
