% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cellexalvr.heatmap.list.R
\docType{methods}
\name{make.cellexalvr.heatmap.list,character-method}
\alias{make.cellexalvr.heatmap.list,character-method}
\title{Create the gene list files for VR}
\usage{
\S4method{make.cellexalvr.heatmap.list}{character}(cvrObj, cellidfile,
  num.sig, outfile, stats_method = NA)
}
\arguments{
\item{cvrObj}{the cellexalObj.RData file}

\item{cellidfile}{file containing cell IDs or the grouping name}

\item{num.sig}{The number of differentials to be returned}

\item{outfile}{The name of the output file}

\item{stats_method}{the stats method to use see getDifferentials() #function definition in file 'getDifferentials.R'}
}
\description{
preload the cellexalObj.RData file
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: cellexalvrR
}}

\keyword{differential,}
\keyword{gene}
\keyword{list}
