% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\name{make.cellexalvr.heatmap}
\alias{make.cellexalvr.heatmap}
\title{Creates a heatmap from a selection of groups}
\usage{
make.cellexalvr.heatmap(cellexalObj, cellidfile, num.sig = 200, outfile)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{num.sig}{The number of differentials to be returned}

\item{outfile}{The name of the output file}
}
\description{
Creates a heatmap from a selection of groups
}
\keyword{heatmap}
