% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\docType{methods}
\name{make.cellexalvr.network}
\alias{make.cellexalvr.network}
\alias{make.cellexalvr.network,cellexalvrR-method}
\title{description of function make.cellexalvr.network}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{outpath}{the outpath}

\item{cutoff.ggm}{The cutoff for the correlation (default = 0.1)}

\item{exprFrac}{which fraction of cells needs to express a gene to be included in the analysis (default 0.1)}

\item{top.n.inter}{get only the n top interations default=130}
}
\description{
Creates a network from selected groups for selected genes
}
\details{
This is the R function that calculates the TF networks shown in the VR environment.
}
\keyword{construction}
\keyword{network}
