% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\docType{methods}
\name{make.cellexalvr.network}
\alias{make.cellexalvr.network}
\alias{make.cellexalvr.network,cellexalvrR-method}
\title{description of function make.cellexalvr.network}
\usage{
make.cellexalvr.network(
  cellexalObj,
  cellidfile,
  outpath,
  cutoff.ggm = 0.1,
  exprFract = 0.1,
  top.n.inter = 130,
  method = c("rho.p", "pcor")
)

\S4method{make.cellexalvr.network}{cellexalvrR}(
  cellexalObj,
  cellidfile,
  outpath,
  cutoff.ggm = 0.1,
  exprFract = 0.1,
  top.n.inter = 130,
  method = c("rho.p", "pcor")
)

\S4method{make.cellexalvr.network}{character}(
  cellexalObj,
  cellidfile,
  outpath,
  cutoff.ggm = 0.1,
  exprFract = 0.1,
  top.n.inter = 130,
  method = c("rho.p", "pcor")
)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{outpath}{the outpath}

\item{cutoff.ggm}{The cutoff for the correlation (default = 0.1)}

\item{top.n.inter}{get only the n top interations default=130}

\item{exprFrac}{which fraction of cells needs to express a gene to be included in the analysis (default 0.01)}
}
\description{
Creates a network from selected groups for selected genes
}
\details{
This is the R function that calculates the TF networks shown in the VR environment.
}
\keyword{construction}
\keyword{network}
