% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\docType{methods}
\name{make.cellexalvr.network}
\alias{make.cellexalvr.network}
\alias{make.cellexalvr.network,cellexalvrR-method}
\title{description of function make.cellexalvr.network}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{outpath}{the outpath}

\item{cutoff.ggm}{The cutoff for the correlation (default = 0.8)}

\item{top.n.inter}{get only the n top interations default=125}
}
\description{
Creates a network from selected groups for selected genes
}
\keyword{construction}
\keyword{network}
