% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrNetworkFunctions.R
\name{make.cellexalvr.network}
\alias{make.cellexalvr.network}
\title{Creates a network from selected groups for selected genes}
\usage{
make.cellexalvr.network(cellexalObj, cellidfile, tf.loc, outfile)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{outfile}{The name of the output file}

\item{location}{of TF file}
}
\description{
Creates a network from selected groups for selected genes
}
\keyword{construction}
\keyword{network}
