% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\docType{methods}
\name{make.cellexalvr.network,character-method}
\alias{make.cellexalvr.network,character-method}
\title{description of function make.cellexalvr.network}
\usage{
\S4method{make.cellexalvr.network}{character}(
  cellexalObj,
  cellidfile,
  outpath,
  cutoff.ggm = 0.1,
  exprFract = 0.1,
  top.n.inter = 130,
  method = c("rho.p", "pcor")
)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{outpath}{the outpath}

\item{cutoff.ggm}{The cutoff for the correlation (default = 0.8)}

\item{top.n.inter}{get only the n top interations default=125}

\item{exprFrac}{which fraction of cells needs to express a gene to be included in the analysis (default 0.1)}
}
\description{
preload the cellexalObj.RData file
}
\section{Functions}{
\itemize{
\item \code{make.cellexalvr.network,character-method}: cellexalvrR
}}

\keyword{construction}
\keyword{network}
