% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCellexalVRObj.R
\docType{methods}
\name{MakeCellexaVRObj}
\alias{MakeCellexaVRObj}
\alias{MakeCellexaVRObj,cellexalvrR-method}
\title{description of function MakeCellexaVRObj}
\arguments{
\item{exdata}{A matrix of expression values (required). Colnames are cell IDs, rownames are unique gene names}

\item{mds.list}{A list of 3-column MDS coords (at least one required)}

\item{specie}{Specify whether data is from mouse of human (required)}

\item{cell.metadata}{Required meta data for cells}

\item{facs}{Surface marker intensities from index sorted cells}

\item{specie}{TEXT MISSING default=c("mouse"}

\item{"human")}{TEXT MISSING default=c("mouse"}

\item{cell.metadata}{TEXT MISSING default=NULL}

\item{facs.data}{TEXT MISSING default=NULL}
}
\description{
Creates a cellexalvr objext with the data given
}
\keyword{heatmap}
