% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_gene_links.R
\docType{methods}
\name{md_gene_links}
\alias{md_gene_links}
\alias{md_gene_links,cellexalvrR-method}
\title{Create a binned annotation column from numeric data}
\usage{
md_gene_links(genes, label = "Click to expand gene list")
}
\arguments{
\item{genes}{a list of strings added to "https://www.genecards.org/cgi-bin/carddisp.pl?gene="}

\item{label}{defaults tp "Click to expand gene list"}
}
\description{
Bin the UMI data into 13 bins for plotting and define a blue <- red color gradient
}
\details{
convert a list of genes into a collapsed markdown text that the user can expand.
}
