% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_gene_links.R
\docType{methods}
\name{md_gene_links}
\alias{md_gene_links}
\alias{md_gene_links,character-method}
\alias{md_gene_links,NULL-method}
\title{log specific formating of a (long) list of strings}
\usage{
md_gene_links(genes, label = "Click to expand gene list")

\S4method{md_gene_links}{character}(genes, label = "Click to expand gene list")

\S4method{md_gene_links}{`NULL`}(genes, label = "Click to expand gene list")
}
\arguments{
\item{genes}{a list of strings added to "https://www.genecards.org/cgi-bin/carddisp.pl?gene="}

\item{label}{defaults to "Click to expand gene list"}
}
\description{
Convert genes to a web link in the log files
}
\details{
convert a list of genes into a collapsed markdown text that the user can expand.
}
