% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNextCoolHeatmapFeature.R
\docType{methods}
\name{myNextCoolHeatmapFeature}
\alias{myNextCoolHeatmapFeature}
\alias{myNextCoolHeatmapFeature,cellexalTime-method}
\title{description of function check}
\usage{
\S4method{myNextCoolHeatmapFeature}{cellexalvrR,character,character}(x, selection, outfile)
}
\arguments{
\item{selection}{the selection file linked to the heatmap}

\item{outfile}{the outfile the VR process is waiting for}

\item{cellexalObj}{the cellexalvrR object}
}
\description{
checks for NA elements in the table and removes them
}
\details{
This function is linked to a button on the VR heatmaps.
In the default cellexalvrR version this function is a blank to allow non VR Bioinformaticians
to start a function they miss and want to implement themselves.
}
