% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyGOIs.R
\docType{methods}
\name{onlyGOIs}
\alias{onlyGOIs}
\alias{onlyGOIs,cellexalvr-method}
\title{description of function onlyGOIs}
\usage{
onlyGOIs(cellexalObj, name)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}

\item{name}{the name of the GIO list (eg TFs or epigenetic)}
}
\description{
Allows the user to select only (G)enes (O)f (I)nterest lists from the object
}
