% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyGOIs.R
\docType{methods}
\name{onlyGOIs}
\alias{onlyGOIs}
\alias{onlyGOIs,cellexalvrR-method}
\title{Non VR method to reduce a cellexalvrR object to only include specififc genes}
\usage{
onlyGOIs(cellexalObj, name)

\S4method{onlyGOIs}{cellexalvrR}(cellexalObj, name)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{name}{the name of the GIO list (eg TFs or epigenetic)}
}
\description{
Allows the user to select only (G)enes (O)f (I)nterest lists from the object
}
\details{
onlyGOIs is able to use the inbuild GOI datasest to subset the main data to e.g. TFs only.

This function is used internally.
}
