% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologyLogPage.R
\docType{methods}
\name{ontologyLogPage}
\alias{ontologyLogPage}
\alias{ontologyLogPage,cellexalvrR-method}
\title{description of function ontologyLogPage}
\usage{
ontologyLogPage(cellexalObj, genes, n, ...)
}
\arguments{
\item{cellexalObj}{the cellexalvrR object}

\item{genes}{a list of gene symbols (IMPORTANT)}

\item{n}{the nth heatmap to report on here}

\item{...}{unused}
}
\description{
creates the GO analysis for a gene list and puts it into the report.
}
