% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{plotDataOnTime}
\alias{plotDataOnTime}
\alias{plotDataOnTime,cellexalTime-method}
\title{description of function plot}
\arguments{
\item{x}{the object}

\item{dat}{the linear data to plot}

\item{color}{the color(s) the data should be plotted in}

\item{ofile}{the figure file (png)}

\item{smooth}{add a smoothed line to the graph (default TRUE)}
}
\description{
plot linear data on the time in a 2d figure
}
