% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{plotDataOnTime_rmd}
\alias{plotDataOnTime_rmd}
\alias{plotDataOnTime_rmd,cellexalTime-method}
\title{create the Rmd figure strings for the plotTime files}
\usage{
plotDataOnTime_rmd(x, dat, color = NULL, ofile, cellexalObj)

\S4method{plotDataOnTime_rmd}{cellexalTime}(x, dat, color = NULL, ofile, cellexalObj)
}
\arguments{
\item{x}{the object}

\item{dat}{the linear data to plot}

\item{color}{the color(s) the data should be plotted in}

\item{ofile}{the figure file (png)}

\item{cellexalObj}{the cellexal object to adjust the path for the markdown file.}
}
\description{
plot linear data on the time in a 2d figure
}
