% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalTime.R
\docType{methods}
\name{plotTime}
\alias{plotTime}
\alias{plotTime,cellexalTime-method}
\title{description of function plot}
\arguments{
\item{x}{the object}
}
\description{
3D plot the time with a line in the time
}
