% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02cellexalTime.R
\docType{methods}
\name{plotTime}
\alias{plotTime}
\alias{plotTime,cellexalLinear-method}
\title{plot the drc with pseudo time colors}
\usage{
plotTime(x)

\S4method{plotTime}{cellexalLinear}(x)
}
\arguments{
\item{x}{the object}
}
\description{
3D plot the cellexalLinear object
}
\details{
Plot the linear data with the correct colors.
Used to debug the linear selections.
}
