% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01cellexalvrClass.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,cellexalvr-method}
\title{description of function print}
\usage{
\S4method{show}{cellexalvrR}(object)
}
\arguments{
\item{x}{the cellexalvr object}

\item{x}{the cellexalvr object}
}
\value{
nothing

nothing
}
\description{
print the cellexalvr

shows the cellexalvr contents
}
