% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderReport.R
\docType{methods}
\name{renderReport}
\alias{renderReport}
\alias{renderReport,cellexalvrR-method}
\title{description of function renderReport}
\usage{
renderReport(cellexalObj)
}
\arguments{
\item{cellexalObj}{the cellexlvrR object}
}
\description{
after one session this function renders the output gitbook
}
