% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderReport.R
\docType{methods}
\name{renderReport}
\alias{renderReport}
\alias{renderReport,cellexalvrR-method}
\alias{renderReport,character-method}
\title{after one session this function renders the output as gitbook}
\usage{
renderReport(cellexalObj)

\S4method{renderReport}{cellexalvrR}(cellexalObj)

\S4method{renderReport}{character}(cellexalObj)
}
\arguments{
\item{cellexalObj}{the cellexlvrR object}
}
\description{
after one session this function renders the output as gitbook
}
\details{
renderReport is the final step to create a html log file.

These log files can be accessed from within the VR environment using the inbuilt browser.
It also creates a zip file that contains all data required to view this report.
}
