% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renew.R
\docType{methods}
\name{renew}
\alias{renew}
\alias{renew,cellexalvrR-method}
\alias{renew,cellexalvr-method}
\title{renew/update a cellexalvrR object structure to the last version}
\usage{
renew(x)

\S4method{renew}{cellexalvrR}(x)

\S4method{renew}{cellexalvr}(x)
}
\arguments{
\item{x}{the object you want to updatae}
}
\description{
updatae the class definition by re-creating the instance This version also makes sure,

that the returned object is an S4 object.
}
\details{
renew is trying to update objects from old versions to the most up to date structure.

It is tested to work on version <= 0.14.5.
}
