% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratExtensions.R
\docType{methods}
\name{run.ddrtree}
\alias{run.ddrtree}
\alias{run.ddrtree,seurat-method}
\title{description of function run.ddrtree}
\arguments{
\item{seuratObj}{A cellexalvr object}
}
\description{
Runs DDRtree for a Seurat class
}
\keyword{ddrtree}
