% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R, R/seuratExtensions.R
\name{run.ddrtree}
\alias{run.ddrtree}
\alias{run.ddrtree}
\title{Runs DDRtree for a Seurat class}
\usage{
run.ddrtree(seuratObj)

run.ddrtree(seuratObj)
}
\arguments{
\item{seuratObj}{A cellexalvr object}

\item{seuratObj}{A cellexalvr object}
}
\description{
Runs DDRtree for a Seurat class

Runs DDRtree for a Seurat class
}
\keyword{ddrtree}
