% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateParts.R
\docType{methods}
\name{savePart}
\alias{savePart}
\alias{savePart,cellexalvrR-method}
\title{description of function savePart}
\usage{
\S4method{savePart}{cellexalvrR}(
  x,
  part = c("meta.cell", "meta.gene", "userGroups", "usedObj", "groupSelectedFrom",
    "colors", "lastGroup"),
  path = NULL
)
}
\arguments{
\item{x}{TEXT MISSING}

\item{part}{TEXT MISSING default= c( 'meta.cell'}

\item{path}{the optional outpath (default x@outpath)}
}
\description{
save only a part of the cellexal object increasing the VR interaction speed
}
\details{
The function savePart is one main part in the load and save speed up implemented for cellexalvrR

Instead of saving all data in each VR script call only the modified data is saved.
This function saves a specific data part of the whole cellexalvrR speeding up the save process.
}
