% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionRegisterGrouping.R
\docType{methods}
\name{sessionRegisterGrouping}
\alias{sessionRegisterGrouping}
\alias{sessionRegisterGrouping,cellexalvrR-method}
\title{description of function sessionRegisterGrouping}
\usage{
sessionRegisterGrouping(x, gName)
}
\arguments{
\item{x}{the cellexalvrR object}

\item{gName}{the group name to register}
}
\description{
Registers a grouping for this session.
}
