% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\docType{methods}
\name{set.specie}
\alias{set.specie}
\alias{set.specie,cellexalvrR-method}
\title{description of function set.specie}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{specie}{The specie required}

\item{specie}{TEXT MISSING default=c("mouse"}

\item{"human")}{TEXT MISSING default=c("mouse"}
}
\description{
Loads TF annotation into cellexalvr object
}
\keyword{TFs}
