% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrFunctions.R
\docType{methods}
\name{set.specie}
\alias{set.specie}
\alias{set.specie,cellexalvrR-method}
\title{Set the specie information in the cellexalvrR object}
\usage{
set.specie(cellexalObj, specie = c("mouse", "human"))

\S4method{set.specie}{cellexalvrR}(cellexalObj, specie = c("mouse", "human"))
}
\arguments{
\item{cellexalObj}{cellexalvr object}

\item{specie}{The specie required options: "mouse" or "human"}
}
\description{
Set the specie information in the cellexalvrR object
}
\details{
It sets the specie slot to the user supplied specie string.
}
\keyword{TFs}
