% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratExtensions.R
\docType{methods}
\name{seurat2cellexalvr}
\alias{seurat2cellexalvr}
\alias{seurat2cellexalvr,seurat-method}
\title{description of function seurat2cellexalvr}
\arguments{
\item{seuratObj}{the suerat object to be converted}
}
\description{
Converts a 'seurat' class to one of 'cellexalvr'. If the cell-cycle score have been calculated

they will be added to the cell.mata table
}
\details{
This function can convert a 'Seurat' (v <3.0.0) object into 'cellexalvrR'.
}
