% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrExportFunctions.R
\name{seurat2cellexalvr}
\alias{seurat2cellexalvr}
\title{Makes the base files needed to run the VR environment from a Seurat object}
\usage{
seurat2cellexalvr(seuratObj)
}
\arguments{
\item{cellexalObj}{A cellexalvr object}
}
\description{
Makes the base files needed to run the VR environment from a Seurat object
}
