% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,cellexalvrR-method}
\title{show the object contents}
\arguments{
\item{object}{the cellexalvr object}
}
\value{
nothing at all
}
\description{
show the cellexalvrR
}
