% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellexalvrClass.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,cellexalvr-method}
\title{description of function show}
\usage{
\S4method{show}{cellexalvr}(object)
}
\arguments{
\item{x}{the cellexalvr object}
}
\value{
nothing
}
\description{
shows the cellexalvr
}
