% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useInbuiltGOIlists.R
\docType{methods}
\name{useInbuiltGOIlists}
\alias{useInbuiltGOIlists}
\alias{useInbuiltGOIlists,cellexalvrR-method}
\title{easily identify e.g. all TFs in the dataset}
\usage{
useInbuiltGOIlists(x, name)

\S4method{useInbuiltGOIlists}{cellexalvrR}(x, name)
}
\arguments{
\item{x, }{TEXT MISSING}

\item{name}{one of ("TFs", 'epigenetic', "CellCycle", "CellSurface")}
}
\description{
An easy function to register the inbuilt (G)enes (O)f (I)nterest lists.
}
