% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userGrouping.R
\docType{methods}
\name{userGrouping}
\alias{userGrouping}
\alias{userGrouping,cellexalvrR-method}
\title{description of function userGrouping}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{cellidfile}{file containing cell IDs}

\item{cellidfile}{TEXT MISSING}
}
\description{
Reads a VR cell selection file and creates a user.grouping column with the information

storing the user defined grouping for later use
}
\details{
VR function that ready the VR grouping information and stores it in the R object.
}
\keyword{userGrouping}
