% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userGrouping.R
\docType{methods}
\name{userGrouping}
\alias{userGrouping}
\alias{userGrouping,cellexalvrR-method}
\title{add a CellexalVR created selection to the cellexalvrR object}
\usage{
userGrouping(cellexalObj, cellidfile)

\S4method{userGrouping}{cellexalvrR}(cellexalObj, cellidfile)
}
\arguments{
\item{cellexalObj, }{cellexalvr object}

\item{cellidfile}{file containing cell IDs}
}
\description{
Reads a VR cell selection file and creates a user.grouping column with the information

storing the user defined grouping for later use
}
\details{
VR function that reads the VR grouping information and stores it in the R object.
Each selection is only added once into the cellexalvrR object.
}
